% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancR_codes}
\alias{cancR_codes}
\title{ICD diagnosis codes}
\format{
A list of diagnosis codes divided into the ICD chapters including a sublist with major diseases only
}
\usage{
cancR_codes
}
\description{
List of diagnosis codes after the ICD chapters
}
\keyword{datasets}
