% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_timR.R
\name{cci_timR}
\alias{cci_timR}
\title{Generation of time-dependent CCI and comorbidities from the Charlson Comorbidity Index}
\usage{
cci_timR(
  data,
  interval = 365.25/2,
  structure = "binned",
  format = "long",
  comorb = F,
  cores = 4
)
}
\arguments{
\item{data}{Full LPR dataset}

\item{interval}{Interval in which the CCIs should be calculated}

\item{structure}{Whether CCI should be the full 18 parametres or binned into 1, 2-3, 4-5 and 6+}

\item{format}{Whether the output should be in long or wide format}

\item{comorb}{Whether the specific comorbidites in the CCI score should be extracted. Default is FALSE.}
}
\value{
cci_full_long/wide: Dates of all CCI scores. Repeted scores in adjacent rows are removed as the status is not changed.
cci_bin_long/wide: Same as cci_all with CCI scores binned 1, 2-3, 4-5 and 6+
comorb_long/wide: Dates of all comorbidites
}
\description{
Generation of time-dependent CCI and comorbidities from the Charlson Comorbidity Index
}
