% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkR.R
\name{checkR}
\alias{checkR}
\title{Detection of positivity violations (empty levels)}
\usage{
checkR(
  data,
  treatment = NULL,
  outcome = NULL,
  vars = NULL,
  id,
  levels = NULL,
  threshold = 0,
  return.counts = F,
  quantiles = "decile"
)
}
\arguments{
\item{data}{data frame to detect positivity violations}

\item{treatment}{treatment stratum that should be included to all covariate combinations (optional)}

\item{outcome}{outcome stratum that should be included to all covariate combinations (optional)}

\item{vars}{vector of covariates to examine for positivity violations}

\item{id}{column indicating unique patient identifier for returning specific NAs}

\item{levels}{the number of covariates for which each treatment and/or outcome level will be counted (default = all covariate combinations)}

\item{quantiles}{quantile argument for categorization of numeric variables. See \code{cutR()} for supported quantiles. Default = "decile"}
}
\value{
prints the variables with positivity violations if present, otherwise none detected.
}
\description{
Detection of positivity violations (empty levels)
}
