% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest.R
\name{closest}
\alias{closest}
\title{Closest}
\usage{
closest(mdose, str, split = T)
}
\arguments{
\item{mdose}{estimated dose in the patient}

\item{str}{strength of the tablet}

\item{split}{whether the tablet can be split}
}
\value{
Returns the dose closest to the estimated
}
\description{
Pick the closest value from a range in vector.
}
