% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closR.R
\name{closR}
\alias{closR}
\title{Pick the closest value from a range in vector.}
\usage{
closR(x, vec = NULL, str, split = T)
}
\arguments{
\item{x}{Input number}

\item{vec}{vector to search for closest value. If NULL, provide str which vec should be created from}

\item{str}{Multiplier for the generic vector}

\item{split}{whether the tablet can be split}
}
\value{
Returns the dose closest to the estimated
}
\description{
Pick the closest value from a range in vector.
}
