% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterci.R
\name{clusterci}
\alias{clusterci}
\title{clusterci}
\usage{
clusterci(
  data,
  timevar,
  event,
  treatment,
  vars,
  cluster,
  time,
  breaks,
  cores = detectCores()/2
)
}
\arguments{
\item{data}{dataset}

\item{timevar}{time-to-event variable}

\item{event}{status indicator as either 0/1 or 0/1/2}

\item{treatment}{treatment variable, must be coded as factor}

\item{vars}{vector of covariates. All variables must be coded as factors and continouous variables must be binned}

\item{cluster}{column indicating cluster variable (e.g. ID)}

\item{time}{time horizon}

\item{breaks}{subdivisions between time = 0 and fu}

\item{cores}{number of cores for parallel processing. Defaults to half of the available cores}
}
\value{
Life_table: Event table \cr
Plot_data: Data for plotting CIF curves \cr
Risks: Absolute risk estimates at the specified time points \cr
Risk_differences: Absolute risk difference at the specified time point \cr
Surv: Indicator of the surv-argument
}
\description{
Wrapper for the binregATE in the mets package for performing causal inference on time-to-event data in a competing risk setting with clustering
}
