% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizR.R
\name{collectR}
\alias{collectR}
\title{Collection of multiple plots into one. Wrapper for ggarrange. Input must be a list of plots.}
\usage{
collectR(plots, collect = T, ...)
}
\arguments{
\item{plots}{List of plots to be collected}

\item{collect}{Whether labels should be collected}

\item{...}{See ggarrange}

\item{nrow}{Number of rows}

\item{ncol}{Number of colums}
}
\value{
Collected plot
}
\description{
Collection of multiple plots into one. Wrapper for ggarrange. Input must be a list of plots.
}
