% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinR.R
\name{combinR}
\alias{combinR}
\title{Generate all possible cominations/permutations}
\usage{
combinR(letters, letters2 = NULL, list = F)
}
\arguments{
\item{letters}{Vector either of length 1 (will be split for each subelement) og length >1}

\item{letters2}{Optional second vector if chunks are to be combined}

\item{list}{Whether all combinations should be returned as a list (useful for looping with lapply)}
}
\value{
Returns all possible combinations of the input vector(s)
}
\description{
All possible combinations of a single vector (ABC: ABC, ACB, BAC, BCA, CAB, CBA) or two separate vectors (AB, CD: ABCD, CDAB)
The function does not allow for replacements.
}
