% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariates_df}
\alias{covariates_df}
\title{covariates_df}
\format{
A data frame with 112257 rows and 21 columns:
\describe{
\item{pnr}{unique patient pnr}
\item{date}{date for the covariate values}
\item{cci}{Charlson Comorbidity Index binned into the groups: 0, 1, 2-3, 4-5 and 6+}
\item{education}{Educational level according to ISCED groups: low, medium and high}
\item{income}{Quantile of the equivalated disposable income: q1, q2, q3 and q4}
\item{marital}{Marital status: Unmarried, married and divorced}
\item{region}{Region of habitance: Capital, central, north, south and zealand}
\item{degurba}{Degree of urbanization in the municipality: city, suburb and rural}
\item{infection, cancer, hema, endo, psych, neuro, cvd, lungs, gi, skin, connective, urinary, congenital}{ICD main groups of covariates: 0 and 1}
...
}
}
\usage{
covariates_df
}
\description{
Simulation of a time-dependent data frame of covariates for the matchR function
}
\keyword{datasets}
