% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizR.R
\name{cpR}
\alias{cpR}
\title{Fix CPR numbers with removed leading zeros}
\usage{
cpR(
  data,
  cpr = cpr,
  extract = FALSE,
  remove.cpr = FALSE,
  return.cpr = FALSE,
  dt = NULL
)
}
\arguments{
\item{data}{dataset}

\item{cpr}{name of the cpr-column}

\item{extract}{TRUE if age and date of birth should be extracted}

\item{remove.cpr}{whether invalid CPRs should be removed, default = F}

\item{return.cpr}{whether the invalid CPRs should be returned as a vector, default = F}

\item{dt}{whether the dataframe should be returned as a data.table}
}
\value{
Returns same dataset with correct CPR numbers and optionally age and date of birth. Invalid CPRs stops the function and returns the invalid CPRs as a vector.
}
\description{
Fix CPR numbers with removed leading zeros
}
