% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrplot.R
\name{crrplot}
\alias{crrplot}
\title{crrplot}
\usage{
crrplot(
  list,
  y = 100,
  col = c("#9B62B8", "#224B87", "#67A8DC", "#D66ACE"),
  outcome = "",
  ylab = "Risk of Event (\%)",
  labpos = rep(0, 4),
  printres = "est",
  labgrp = grps,
  fu = 120,
  breaks = 12,
  unit = "Years",
  respos.x = 0,
  respos.y = 0,
  ylab.pos = 0,
  legend.pos = c(0.5, 0.95),
  tscale = 1,
  ndigits_est = 1,
  ndigits_res = 1,
  censur = F
)
}
\arguments{
\item{list}{crrlist or plist object from the crrstrat function}

\item{y}{Value of y-axis in whole number (eg. 50)}

\item{col}{color palette as a vector}

\item{outcome}{Labels for outcome}

\item{ylab}{Label for y-axis}

\item{labpos}{adjustments for point estimates on the plot. Can be removed with NA}

\item{printres}{Which results should be printed. Estimates="est", hazard ratio = "hr", risk difference = "rd" and risk ratio = "rr"}

\item{labgrp}{group labels}

\item{fu}{length of follow-up (default = 120 months)}

\item{breaks}{the subdivisions between time = 0 and follow-up length (defaults to 12)}

\item{unit}{Whether the time unit is years, months or days (defaults to "Years")}

\item{respos.x}{shift of the result label on the x-axis}

\item{respos.y}{shift of the result label on the y-axis}

\item{ylab.pos}{shifting of y-axis label position}

\item{legend.pos}{legend position as vector}

\item{tscale}{Global scaling parameter if text needs to change size}

\item{ndigits_est}{No of digitis in the estimates}

\item{ndigits_res}{No of digits in the results}

\item{censur}{Censuring of counts < 5. Default = FALSE}
}
\value{
Adjusted Kaplan-Meier or cumulative incidence curves
}
\description{
Automatic plot of adjusted Kaplan-Meier and cumulative incidence curves
}
