% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crrstrat.R
\name{crrstrat}
\alias{crrstrat}
\title{Competing risk regression outcomes and plot-preparation}
\usage{
crrstrat(
  data,
  timevar,
  event,
  group,
  surv,
  multi = FALSE,
  quantiles = T,
  type,
  vars,
  form,
  timepoints = seq(6, 120, 6),
  t2 = 120,
  printres = c(seq(0, 60, 12), 120)
)
}
\arguments{
\item{data}{Input dataframe}

\item{timevar}{Time-to-event column}

\item{event}{Event indicator. In case of competing risk the structure should be censoring=0, event=1 and death=2. Otherwise
0/1 can be used in case of no competing risks.}

\item{group}{Grouping variable, between which the contrasts are estimated}

\item{surv}{If competing risks = FALSE, otherwise = TRUE.}

\item{multi}{If more than 2 groups are assigned to the "group" argument.}

\item{quantiles}{If median time to event should be returned.}

\item{type}{Assign covariates for adjustment.
uni: No adjustment other than the "group" variable
age-sex: Age and sex standardization
select: Uses "vars" to select variables that should be adjusted for.
custom: The right-hand side of the formula can be customized as a character string.}

\item{vars}{Only applicable if "type" = "select".}

\item{form}{Only applicable if "type" = "custom".}

\item{timepoints}{Defines timepoints of interest at which estimates should be estimated. Defaults to 6 month intervals}

\item{t2}{Time-horizon (maximum time point to estimate). Defaults to 120 (10 years).}

\item{printres}{Defines which time point esimates should be reported.}
}
\value{
Life_table: Event table
Plot_data: Data for plotting CIF curves
HR: Hazard ratio between the groups
Models: Model objects (cause1 and cause2)
diag: Diagnostics for assessing proportionality
Risks: Absolute risk estimates at the specified time points
Risk_differences: Absolute risk differences at the specified time points
Risk_ratios: Absolute risk ratios at the specified time points
Event_proportions: "Before" estimates the risk of event within a certain timepoint (e.g. x\% of the events occurred within).
"After" estimates the risk of residual events between timepoint x and ten years.
"Window" estimates the percentage of events within six month windows (e.g. x\% of the events occurred between t1 and t2)
Conditional: Conditional risk estimates
Surv: Indicator of the surv-argument
}
\description{
The function uses stratified cause-specific Cox regression and G-computation to estimate the risk of event at a specific
timepoint taking competing risks into account. The function also provides pointwise risk differences and ratios and prepares
the subsequent plotting of the cumulative incidence curves.
The function is a wrapper of the ate-function from riskRegression by T. Gerds et al.
}
