% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{Data summary}
\usage{
data_summary(data)
}
\arguments{
\item{data}{Dataset}
}
\value{
Returns bar charts or density plots depending on format. Numerical variables with less than 5 unique values are considered as factors.
}
\description{
Graphical overview of entire dataset
}
