% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defusR.R
\name{defusR}
\alias{defusR}
\title{Defuse arguments to character string regardless of quoted or non-quoted in functions.}
\usage{
defusR(input, data = NULL)
}
\arguments{
\item{input}{quoted or non-quoted vector (e.g. variable names)}

\item{data}{data frame if tidyselection is used.}
}
\value{
the defused argument as a character string of class "def_ex" for later detection
}
\description{
Defuse arguments to character string regardless of quoted or non-quoted in functions.
}
