% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatR.multi.R
\name{estimatR.multi}
\alias{estimatR.multi}
\title{estimatR.multi}
\usage{
estimatR.multi(data, timevar, event, group, names, ...)
}
\arguments{
\item{data}{dataset}

\item{timevar}{Character vector of time variables. If missing "t_" is assumed to be prefix for all names in the "events" vector}

\item{event}{Character vector of event variables}

\item{group}{Character vector of grouping variables}

\item{names}{Element names of the returned list of models. If missing the "events" names are used.}

\item{...}{See arguments in estimatR()}
}
\value{
A named list of models with the estimatR function
}
\description{
Convenience function for performing multiple estimatR functions (e.g. multiple outcomes or risk factor). Each argument is recycled to avoid repeated names.
}
