% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatR.R
\name{estimatR}
\alias{estimatR}
\title{Absolute risk estimation of time-to-event data with competing risks}
\usage{
estimatR(
  data,
  timevar,
  event,
  group,
  survscale = "AM",
  method = "cox",
  vars,
  event.form = NULL,
  time = 120,
  breaks = 12,
  cause = 1,
  pl = T,
  digits = 4,
  event.digits = 2,
  alpha = 0.05,
  survtime = T,
  proportions = F,
  conditional = F,
  diagnostics = F,
  plot = T,
  verbose = T
)
}
\arguments{
\item{data}{dataset}

\item{timevar}{Time variable}

\item{event}{Status indicator}

\item{group}{Grouping variable}

\item{survscale}{Whether overall survial should be estimated as survival or all-cause mortality (1-survival)}

\item{vars}{Only applicable when "select" is chosen as type. The variables are added to the model as an additive model}

\item{time}{Time horizon of interest. Defaults to 60 (e.g. 5-years)}

\item{breaks}{Interim time points of interest. Defaults to 12 months (1-year gaps)}

\item{cause}{cause of interest, default = 1}

\item{pl}{Whether product.limit in ATE shoulde be T or F (default = T)}

\item{digits}{number of digits risk estimates in the returned results}

\item{event.digits}{Rounding of event times. Default is 2 to preserve exact times}

\item{alpha}{alpha level for the estimation of confidence intervals and p-values. Default = 0.05}

\item{survtime}{Whether median time to event should be calculated (default = T)}

\item{proportions}{Whether risk of event in different windows should be estimated (default = F)}

\item{conditional}{Whether conditional risk at the time horizon should be calculated (default = F)}

\item{diagnostics}{whether Scaled Schoenfield residuals should be visualized (default = F)}

\item{plot}{whether estimates for plotR should be performed (default = T)}

\item{verbose}{whether status should be printed to the console (default = T)}

\item{type}{Model specification, Can be univariate ("uni"), Age- and sex standardized ("age-sex"), Multivariate with variable selection ("select"). In this case vars should be a vector of the covariates. "custom" allows for free modelling where the form-argument contains the formula.}

\item{form}{Only applicable when "custom" is chosen as type. Free specification of the model as the right-hand side of the formula.}
}
\value{
List of class "estimatR" containing the following:
table: Event table \cr
models: Model objects (cause1 and cause2) \cr
risks: Absolute risk estimates at the specified time points \cr
plot_data: Data for plotting CIF curves \cr
time_to_event: Median survival time \cr
hr: Hazard ratio between the groups \cr
difference: Absolute risk difference at the specified time horizons \cr
ratio: Absolute risk ratios at the the specified time horizon \cr
counts: Event and group counts in the contrasted groups \cr
info: information on arguments for extraction
Optional:
diag: Diagnostics for assessing proportionality \cr
proportions: \cr
"Before" estimates the risk of event within a certain timepoint (e.g. x\% of the events occurred within). \cr
"After" estimates the risk of residual events between timepoint x and ten years. \cr
"Window" estimates the percentage of events within six month windows (e.g. x\% of the events occurred between t1 and t2) \cr
conditional: Conditional risk estimates \cr
}
\description{
Absolute risk estimation and comparison of two or more groups with a time-to-event outcome.
Wrapper for the prodlim and ate-functions in riskRegression.
}
\examples{
estimatR(analysis_df, ttt, event, X1, type = "select", vars = c(X6,X7))


}
