% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_struct.R
\name{event_struct}
\alias{event_struct}
\title{Event structuring}
\usage{
event_struct(
  data,
  index,
  fu,
  outcomes,
  competing,
  pattern = "_date",
  unit = "months",
  keep_dates = F
)
}
\arguments{
\item{data}{Dataframe}

\item{index}{Index time point 'T=0' provided in date format}

\item{fu}{End of follow-up or time of death in date format}

\item{outcomes}{Vector of single or multiple column names with the event of interest in date format}

\item{competing}{Single column indicating either status or dates of the competing event. If passed as a status (0/1) the dates are assumed to stem from follow-up (e.g. death)}

\item{pattern}{Indicates the name pattern of the outcomes such as index_event or event_date}

\item{unit}{Whether time-to-event should be reported in months or years.}

\item{keep_dates}{Whether the original event dates should be kept.}
}
\value{
For each outcome, a status indicator of 0/1/2 and a time-to-event column are returned. If competing is missing, the levels are 0/1
For the competing event, a status indicator of 0/1 and a time-to-event is returned.
For patients without an event the function returns the time to last follow-up if status = 0 and time to death if status = 2
}
\description{
Converts event dates to status indicator and a time-to-event variable.
The status indicator will be assigned to:
0 if alive and event-free,
1 in the presence of the event of interest
2 in the presence of the competing event, often death
}
