% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_match.R
\name{ex_match}
\alias{ex_match}
\title{ex_match}
\usage{
ex_match(
  data,
  pnr,
  case,
  index,
  fu,
  cat,
  dat,
  exclude,
  n_controls,
  seed = 1,
  cores = 4
)
}
\arguments{
\item{data}{data frame with cases and controls defined}

\item{pnr}{pnr}

\item{case}{1/0}

\item{index}{index of cases, controls should be blank}

\item{fu}{follow-up time for all patients}

\item{cat}{categorical matching parametres}

\item{dat}{time-dependent matching paramtres}

\item{exclude}{exclusion of event before index}

\item{n_controls}{number of desired controls}

\item{seed}{seed}

\item{cores}{default = 4}
}
\value{
returns the original data frame with n_controls matches per case with set indicating the risk set.
}
\description{
Time-dependent exact matching for matched cohort studies.
}
