% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractR.R
\name{extractR}
\alias{extractR}
\title{extractR}
\usage{
extractR(list, vars = c("counts", "risks", "diff"), shape = "long")
}
\arguments{
\item{list}{estimatR object}

\item{vars}{Which outcome measures to provide (counts, risks, differneces and/or ratios)}

\item{shape}{Format of the table (long or wide).}
}
\value{
Returns a data frame compatible with flextable
}
\description{
Automatic extraction of key results from the estimatR function
}
