% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatR.R
\name{formatR}
\alias{formatR}
\title{formatR}
\usage{
formatR(
  data,
  cat_vars = c(case, sex, cci, region, education, income, marital),
  num_vars = c(period, age_group),
  cut_vars = c(age, index),
  seqlist = list(),
  labels = list(),
  lab_to_lev = F,
  names = list(),
  layout = NULL
)
}
\arguments{
\item{data}{dataframe. Works with piping}

\item{cat_vars}{Vector of categorical covariates which should be formatted to factors with labels}

\item{num_vars}{Vector of numerical covariates that should be cut}

\item{cut_vars}{vector of vars to slice}

\item{case.labs}{Labels for the case/control column}
}
\value{
Returns the same dataframe with print-friendly labels for the typical full matching setup.
}
\description{
formatR
}
