% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupR.R
\name{groupR}
\alias{groupR}
\title{Extract first n groups in data frame}
\usage{
groupR(data, grps, n)
}
\arguments{
\item{data}{dataframe}

\item{grps}{grouping variable}

\item{n}{number of groups to extract}
}
\value{
filtered dataset including only n first groups
}
\description{
Extract first n groups in data frame
}
