% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includR.R
\name{includR}
\alias{includR}
\title{Inclusion criteria for registry-studies}
\usage{
includR(
  data,
  exclusion.ex,
  exclusion.out,
  subgroup = NULL,
  age.limit = NULL,
  period = NULL,
  look.back = 365.25 * 5,
  index.shift = 0,
  fu = "fu",
  birth = "birth",
  remove = NULL,
  export = F,
  dt = F
)
}
\arguments{
\item{data}{joined data frame}

\item{exclusion.ex}{vector of exposure variables that should not be present before index}

\item{exclusion.out}{vector of outcome variables that should not be present before index}

\item{subgroup}{filter for subgroups in the format: "level \%in\% c('a', 'b')" Can also be applied as custom filter}

\item{age.limit}{age limit for inclusion as numeric value}

\item{period}{character vector of length 2 specifying the enrollement period in the format c("1980-01-01", "2022-01-01")}

\item{look.back}{look back in days period where exposure and outcomes are not allowed to happen (default is five years - 365.25*5)}

\item{index.shift}{number of days before index where exposure and outcomes are allowed, default = 0}

\item{fu}{name of the follow-up date (default: fu)}

\item{birth}{name of the birth date variable (default: birth)}

\item{export}{whether a data frame of the flowchart should be exported}
}
\value{
A data frame with filtered cases and all controls + an rds file for flow chart
}
\description{
Inclusion criteria for registry-studies
}
