% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencR.R
\name{inferencR}
\alias{inferencR}
\title{Causal inference of time-to-event data}
\usage{
inferencR(
  data,
  treatment,
  timevar,
  event,
  vars,
  outcome.vars,
  time = 120,
  breaks = 12,
  estimator = "AIPTW",
  survtime = T,
  survscale = "AM",
  bins = 0.5,
  digits = 4,
  event.digits = 2
)
}
\arguments{
\item{data}{dataframe}

\item{treatment}{column indicating treatment variable. Currently only binary treatments are supported.}

\item{timevar}{Time variable}

\item{event}{Status indicator}

\item{vars}{variables that the treatment, censoring and event models should include}

\item{outcome.vars}{variables that should only be included in the outcome models}

\item{time}{Time horizon of interest. Defaults to 60 (e.g. 5-years)}

\item{breaks}{Interim time points of interest. Defaults to 12 months (1-year gaps)}

\item{estimator}{whether the estimator should be "IPTW", "GFORMULA" or "AIPTW" (default)}

\item{survtime}{Whether median time to event should be calculated (default = F)}

\item{survscale}{Whether overall survial should be estimated as survival or all-cause mortality (1-survival)}

\item{bins}{bins for the weights plot, default = 0.5}

\item{digits}{for rounding of eventtimes}

\item{event.digits}{whether eventtimes should be rounded. Default is 2 to preserve exact times}
}
\value{
time_to_event: Median survival time \cr
table: Event table \cr
plot_data: Data for plotting CIF curves \cr
models: Model objects (cause1 and cause2) \cr
diag: Diagnostics for assessing proportionality \cr
risks: Absolute risk estimates at the specified time points \cr
differences: Absolute risk difference at the specified time horizons \cr
ratios: Absolute risk ratios at the the specified time horizon \cr
counts: Event and group counts in the contrasted groups \cr
info: information on arguments for extraction
}
\description{
Average treatment effect of point intervention. Wrapper for the ATE function in riskRegression
}
