% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_regs.R
\name{load_regs}
\alias{load_regs}
\title{load_regs}
\usage{
load_regs(
  regs = c("lpr", "pop", "cancer", "lmdb", "opr"),
  keep = NULL,
  n = NULL,
  id_filter = NULL,
  pattern = NULL
)
}
\arguments{
\item{regs}{which registers should be loaded. Default is all (lpr, pop, pato, cancer, lmdb and opr)}

\item{keep}{which variables should be kept provided as a list("lpr" = c("vars")).}

\item{n}{number of observations that should be loaded}

\item{id_filter}{optional possibility to limit the registers to a defined patient population of PNRs}

\item{pattern}{regex patterns provided as a list("lpr" = "DC92|DC21")}
}
\value{
Returns the selected registers to the global environment
}
\description{
Simple loading function of the most used registers including variable selection, regex filtering and id_list filtering.
}
