% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_drugs.R
\name{map_drugs}
\alias{map_drugs}
\title{Map_drugs}
\usage{
map_drugs(
  data,
  drug,
  pnr = pnr,
  atc = atc,
  eksd = eksd,
  apk = apk,
  strnum = strnum,
  packsize = packsize,
  split = T
)
}
\arguments{
\item{data}{Dataframe, typically LMDB}

\item{drug}{Regex filtering drugs of interest (e.g. L04AD01 or L0\link{34})}

\item{pnr}{pnr}

\item{atc}{atc}

\item{eksd}{eksd}

\item{apk}{apk}

\item{strnum}{strnum}

\item{packsize}{packsize}

\item{split}{whether the tablets can be split or dose every second day is an option}
}
\value{
A dataframe with dose, duration, cumulated dose as time-varying measures.
}
\description{
Dynamic mapping of drug regimens in the National Danish Pharmaceutical Register.
}
