% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{match_df}
\alias{match_df}
\title{match_df}
\format{
A data frame with 40,100 rows and 11 columns:
\describe{
\item{pnr}{unique patient pnr}
\item{case}{whether case = 1 or control = 0}
\item{index}{index date for cases. NA for controls}
\item{follow}{date of last follow-up which can be death, emigration or end of database}
\item{birth}{date of birth}
\item{sex}{patient sex, f = female, m = male}
\item{skinc}{date of first skin cancer occurence}
\item{imm_sup}{index of first diagnosis of immunosuppression}
\item{random1, random2}{random variables}

...
}
}
\usage{
match_df
}
\description{
Simulation of a population data frame before using the matchR function
}
\keyword{datasets}
