% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missR.R
\name{missR}
\alias{missR}
\title{Overview of NAs in a dataframe}
\usage{
missR(
  data,
  vars,
  drop.rows = F,
  drop.cols = F,
  return.id = F,
  na.remove = "all.na",
  return.data = F,
  dt = NULL,
  print = T,
  verbose = T
)
}
\arguments{
\item{data}{data frame}

\item{vars}{vars where the na check should be beformed. If missing the whole data frame is analysed}

\item{drop.rows}{whether to remove rows containing all NA values, default = F}

\item{drop.cols}{whether to remove columns contain all NA values, default = F}

\item{return.id}{whether rows with any NA should be returned, default = F}

\item{na.remove}{whether all ("all.na") or any ("any.na") NAs should be removed (default = "all.na")}

\item{dt}{whether the data.frame should be returned as a data.table, default = F}

\item{print}{whether the NA check should be printed in the console, default = T}

\item{verbose}{whether cli messages should be printed, default = T}
}
\value{
Prints whether any NAs are detected and returns a data frame with IDs and columns with NA
}
\description{
Overview of NAs in a dataframe
}
\details{
If drop.cols or drop.rows are TRUE, the data.frame is returned as modified. Otherwise the table of missing data is returned.
}
