% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x, na.rm = T)
}
\arguments{
\item{x}{Vector of values}

\item{na.rm}{Whether NAs should be removed, defaults to TRUE}
}
\value{
The most common value of the vector excluding NAs
}
\description{
Get the mode (most common value) of a vector.
}
\examples{
mode(c(1,1,2,3))
}
