% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizR.R
\name{multitaskR}
\alias{multitaskR}
\title{Start a multisession with automatic reset}
\usage{
multitaskR(cores, gb = NULL)
}
\arguments{
\item{cores}{number of cores/workers}
}
\value{
starts a multisession and reverts to sequential plan on.exit of the parent function (outer function)
}
\description{
Start a multisession with automatic reset
}
