% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngroups.R
\name{n_groups}
\alias{n_groups}
\title{ngroups}
\usage{
n_groups(data, grps, n)
}
\arguments{
\item{data}{dataframe}

\item{grps}{grouping variable}

\item{n}{number of groups to extract}
}
\value{
filtered dataset including only n first groups
}
\description{
Extract first n groups in data frame
}
