% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizR.R
\name{numbR}
\alias{numbR}
\title{Format numeric vectors}
\usage{
numbR(numbers, digits = 1, nsmall, ama = F, trim = T, sign = F)
}
\arguments{
\item{numbers}{numeric value or vector for formatting}

\item{digits}{number of digits}

\item{nsmall}{number of zero-digits}

\item{ama}{whether the numbers should be printed according to AMA guidelines (no digits on values >= 10). Default = F.}

\item{trim}{whether the leading padding should be trimmed (default = T)}

\item{sign}{whether the number should be rounded using significant figures specified using the digits argument (defualt = F)}
}
\value{
returns a vector of same lentgh with formatted digits
}
\description{
Format numeric vectors
}
\examples{
numbR(c(5,2,4,10,100, 41.2), ama=F)
numbR(c(5,2,4,10,100, 41.2), ama=T)

}
