% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfun.R
\name{pfun}
\alias{pfun}
\title{P-function}
\usage{
pfun(x)
}
\arguments{
\item{x}{A p-value}
}
\value{
Prints the raw p-value according to AMA manual of style
}
\description{
The function takes a raw p-value and prints it according to the AMA manual of style
}
\examples{
pfun(0.0245)
pfun(0.000001)
pfun(0.048)
pfun(0.056)
}
