% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_denmark}
\alias{population_denmark}
\title{population_denmark}
\format{
A data frame with 1296 rows and 4 columns:
\describe{
\item{sex}{sex, m=male, f=female}
\item{age_group}{age groups in 5-year intervals}
\item{year}{year}
\item{population}{number of inhabitants in Denmark in the given stratum}

...
}
}
\usage{
population_denmark
}
\description{
Population tables for Denmark stratified on 5-year age groups and sex in the period 1990-2025
}
\keyword{datasets}
