% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population_who}
\alias{population_who}
\title{population_who}
\format{
A data frame with 36 rows and 3 columns:
\describe{
\item{sex}{sex, m=male, f=female}
\item{age_group}{age groups in 5-year intervals}
\item{population}{number of inhabitants in Denmark in the given stratum}

...
}
}
\usage{
population_who
}
\description{
WHO standard population in 5-year age groups and sex
}
\keyword{datasets}
