% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positivity.R
\name{positivity}
\alias{positivity}
\title{Positivity}
\usage{
positivity(data, treatment, vars, id = id)
}
\arguments{
\item{data}{data frame to detect positivity violations}

\item{treatment}{the main stratum which all covariates should include all levels of}

\item{vars}{the covariates to examine for positivity violations. State multiple variables as c(var1, var2, var3) without quotes.}
}
\value{
prints the variables with positivity violations if present. Otherwise none detected. Also returns as either NULL or character for downstream use.
}
\description{
Detection of positivity violations (empty levels)
}
