% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powershell.R
\name{powershell}
\alias{powershell}
\title{powershell}
\usage{
powershell(text)
}
\arguments{
\item{text}{R script that needs to be converted to powershell format}
}
\value{
Printed text compatible with powershell
}
\description{
Converter from Rscript to powershell format
}
\examples{
powershell('rscript')

}
