% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstrat.R
\name{pstrat}
\alias{pstrat}
\title{Life-table}
\usage{
pstrat(data, timevar, event, group, survscale = "OS", horizon = 120)
}
\arguments{
\item{data}{dataframe}

\item{timevar}{time to event column}

\item{event}{event column with either 0/1 structure or 0/1/2 in case of competing risks}

\item{group}{optional grouping column}

\item{survscale}{Whether estimates should be presented as Overall survival or All-Cause Mortality}

\item{horizon}{time horizon (default=120 months)}
}
\value{
Life_table: Event table \cr
Plot_data: Data for plotting CIF curves \cr
Median_surv: Median survival time \cr
Surv: Indicator for the type of time-to-event analysis
}
\description{
Convenience wrapper for the prodlim function in riskRegression
}
