% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratR.R
\name{ratR}
\alias{ratR}
\title{Directly standardized incidence rates using the WHO standard population}
\usage{
ratR(
  data,
  group,
  strata = list(c("year")),
  pyears = 1e+05,
  ci.method = "lognormal",
  index,
  age = age,
  sex = sex
)
}
\arguments{
\item{data}{data set containing age, sex and index date and group}

\item{group}{optional if incidence rates should be provided per group}

\item{strata}{list of vectors for which strata the incidence rates should be reported (e.g. per age-group and sex)}

\item{pyears}{the unit of the incidence rate. Default is per million person years}

\item{ci.method}{the method for derivation of confidence intervals. Default is "lognormal"}

\item{index}{variable name of the index data}

\item{age}{name of the age variable}

\item{sex}{name of the sex variable}
}
\value{
a standardized incidence rate of the overall population and, if specified in the strata-argument, stratified incidence rates
}
\description{
Directly standardized incidence rates inspired by the dsr package by Matt Kumar. The incidence rates are standardized on 5-year age intervals and sex for each year in the period 1990-2025 based on population tables of the Danish population. Lastly the incidence rates are weighted based on the WHO world population.
}
