% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readR.R
\name{readR}
\alias{readR}
\title{Load csv, excel, rds and parquet files}
\usage{
readR(path, leading.zeros = T, na = "", ...)
}
\arguments{
\item{path}{path for the file to load.}

\item{leading.zeros}{whether leading zeros should be kept (default = T)}

\item{na}{character vector specifying NA strings (default = "")}

\item{...}{arguments passes to subfunctions}
}
\value{
the given path imported as a data frame
}
\description{
Wrapper for the fread, readxl, readRDS and read_parquet functions with automatic detection of file extension.
}
