% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{redcap_df}
\alias{redcap_df}
\title{Simulated Redcap dataset}
\format{
A data frame with 500 rows and 15 columns:
\describe{
\item{id}{unique patient pnr}
\item{sex}{patient sex, 1 = female, 2 = male}
\item{birth}{date of birth}
\item{age}{age at surgery in years}
\item{followup}{date of last follow-up which can be death, emigration or end of database}
\item{date_of_surgery}{date of surgery}
\item{size}{size of primary tumor in mm}
\item{type}{type of tumor, 0 = benign, 1 = in situ, 2 = malignant}
\item{localisation}{localisation of the primary tumor}
\item{necrosis}{whether necrosis was present in the primary tumor}
\item{margins}{whether there were positive or negative margins at last surgery}
\item{cd10, sox10, ck}{presence or abscence of immunohistochemical markers}
\item{death_date}{date of death}
\item{recurrence_date}{date of first recurrence}
\item{metastasis_date}{date of first metastasis}
...
}
}
\usage{
redcap_df
}
\description{
Simulation of a redcap data frame for basic data management
}
\keyword{datasets}
