% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportR.R
\name{reportR}
\alias{reportR}
\title{reportR}
\usage{
reportR(
  data,
  casename,
  vars = c(period, age_group, sex, education, income, cci, region, marital),
  table = F,
  plot = F,
  type = "simple",
  cols = cancR_palette,
  headings = list(),
  ...
)
}
\arguments{
\item{data}{matched dataset}

\item{casename}{name or number of cases (e.g. "1" or "CLL")}

\item{vars}{vars that should be in the table}

\item{table}{whether a table should be made (default = F)}

\item{plot}{whether a plot should be made (default = F)}

\item{headings}{List specifying labels for variable names}

\item{...}{passed to tablR}
}
\value{
Prints the matching report, table and plot. Returns af list of the table and plot.
}
\description{
Automatic tablR function for overview of matched and unmatched cases
}
