% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_id.R
\name{rolling_id}
\alias{rolling_id}
\title{rolling_id}
\usage{
rolling_id(data, id = id, lab = order)
}
\arguments{
\item{data}{dataset}

\item{id}{column which unique ids should be based on}

\item{lab}{label for new unique id column}
}
\value{
adds a new column to the dataset with unique ids based on original id conditional on a sorting
}
\description{
Assigns rolling ID. Useful after sorting a dataset, which changes the ID order. Works with piping
}
