% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollR.R
\name{rollR}
\alias{rollR}
\title{Perform rolling operations}
\usage{
rollR(
  data,
  by,
  order,
  sort = 1L,
  label = grp,
  type = "roll",
  dt = F,
  vars,
  interval,
  lag = 1
)
}
\arguments{
\item{data}{dataset}

\item{by}{vector of grouping column labels}

\item{order}{vector of variables to order the dataset}

\item{sort}{vector of the values c(1, -1) of length equal to the order argument}

\item{label}{label for new unique id column}

\item{type}{type of grouping operation, see details. Default = "roll"}

\item{dt}{whether a data.table should be returned}

\item{vars}{the variable used for computing lagged differences if type = "interval}

\item{interval}{vector of length 2 for evaluation of whether the lagged differences are within the two bounds}

\item{lag}{length of lag for lagged differences (default = 1)}
}
\value{
adds a new column to the dataset with unique ids based on original id conditional on a sorting
}
\description{
Useful for conditional grouping
}
\details{
Types of roll include:
\itemize{
\item roll: Scans the by argument of unique values and assigns these ids
\item count: simple row counter by group
\item interval: groups based on whether the lagged difference is within the bounds of an interval (arguments vars, interval and)
}
}
