% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savR.R
\name{savR}
\alias{savR}
\title{Save plots and tables}
\usage{
savR(
  object,
  name = NULL,
  width = 154,
  height,
  unit = "mm",
  scale = 2,
  dpi = 900,
  device = NULL,
  compression = "lzw",
  format = NULL,
  parquet.format = "zstd",
  parquet.compression = 19,
  size = 9,
  table.width = 1,
  folder = NULL,
  sep = ";",
  verbose = T
)
}
\arguments{
\item{object}{object to save. If object = "session" the sessionInfo will be exported as a flextable.}

\item{name}{File name of saved object without extension}

\item{width}{width}

\item{height}{heigth. If missing autoscaling is performed}

\item{unit}{mm or cm}

\item{scale}{to fit}

\item{dpi}{resolution}

\item{device}{cairo for special occasions}

\item{compression}{for tiff}

\item{format}{choose between pdf, svg, tiff, jpg and png}

\item{parquet.format}{format for parquet files, default = "zstd"}

\item{parquet.compression}{compression amount for parquet files, default = 19}

\item{size}{text size for flextables, default = 9}

\item{table.width}{table width for flextables, default = 1 (full width)}

\item{folder}{optional subfolder location, default is working directory}

\item{sep}{separator for csv files, default = ";"}
}
\value{
Saves object automatically in current project folder
}
\description{
Wrapper for ggsave with default settings and automatic saving of multiple formats
}
