% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchR.R
\name{searchR}
\alias{searchR}
\title{Find covariates or outcomes from the registers}
\usage{
searchR(
  reglist,
  search.list,
  name.list = NULL,
  sub.list = list(),
  sub.labels = NULL,
  exclusion.list = list(),
  slice = "first",
  format = "date",
  date.filter = NULL,
  register,
  match = "start",
  casename = "index",
  pnr = "pnr",
  cores = 4,
  dt = F
)
}
\arguments{
\item{reglist}{list of dataframe(s)}

\item{search.list}{list with list(label = diagnosis code) structure.}

\item{name.list}{list of labels for the main columns (e.g. list("case" = "diabetes))}

\item{sub.list}{list of variables where the diagnosis code should be kept (e.g. list("KOL" = "subtype"))}

\item{sub.labels}{list of labels for the diagnosis code colum (e.g. list("subtype" = list("a" = c("DC1", "DC2"))))}

\item{exclusion.list}{list with same structure as search.list for exclusion codes}

\item{slice}{which rows should be selected (first(default)/last/all)}

\item{format}{whether selected rows should contain date or 1}

\item{date.filter}{character date for filtering rows before this date}

\item{register}{specification of register if input is a single data.frame}

\item{match}{the match of the regex code (match, start, end or contains(default))}

\item{casename}{the name of the case variable (default = "index")}

\item{pnr}{name of the pnr column}
}
\value{
A dataframe with the selected columns with corresponding diagnosis codes
}
\description{
Find covariates or outcomes from the registers
}
