% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatR.R
\name{simulatR}
\alias{simulatR}
\title{Function for simulating data}
\usage{
simulatR(
  simulate,
  n = 10,
  start.date = "2000-01-01",
  end.date = "2025-12-31",
  pattern.list = list(),
  lpr.diag.count = 5,
  lmdb.max.prescriptions = 20,
  lmdb.max.packages = 3,
  opr.diag.count = 5,
  pop.min.age = 20,
  pop.max.age = 100,
  pop.sex = 0.5,
  pop.mortality = 0.1,
  match.cases = n * 0.1,
  match.birth = "1960-01-01",
  match.index = c("1990-01-01", "2010-01-01"),
  covariates.period = c("1980-01-01", "2025-01-01"),
  format = "wide",
  seed = 1
)
}
\arguments{
\item{simulate}{Character vector of what to simulate (see details).}

\item{n}{number of unique pnrs to simulate}

\item{start.date}{starting date of the register/dates}

\item{end.date}{end.date of the register/dates}

\item{pattern.list}{list of vectors of diagnoses codes for each register in the format ("lpr" = c("DC92", "DC21"))}

\item{lpr.diag.count}{maximum number of diagnoses per patient in LPR}

\item{lmdb.max.prescriptions}{maximum number of prescriptions per patient in LMDB}

\item{lmdb.max.packages}{maximum number of packages purchased per day}

\item{opr.diag.count}{maximum number of diagnoses per patient in OPR}

\item{pop.min.age}{minmmum age in the population}

\item{pop.max.age}{maximum age in the population}

\item{pop.sex}{distribution of sex in the population, default = 0.5}

\item{pop.mortality}{mortality rate in the population, default = 0.1}

\item{match.cases}{number of cases for the "match" dataset}

\item{match.birth}{maximum birth date. Cohorts are established backwards based on n assuming 100.000 per cohort}

\item{match.index}{vector of length 2 with range of dates for indices}

\item{covariates.period}{vector of length 2 withrange of dates for covariate status}

\item{seed}{for reproducibility}
}
\value{
a single data frame or named list of data frames with simulated registers
}
\description{
Includes the simulation functions in the heaven package.
}
\details{
The following types of data can be simulated:
\itemize{
\item Dates (\code{"dates"})
\item Registers (lpr, lmdb, opr, immune, pop, pato, match, covariates)
}
}
