% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_drugs.R
\name{sum_drugs}
\alias{sum_drugs}
\title{Drug summary}
\usage{
sum_drugs(mobj, pnr = pnr, trend = F, scale)
}
\arguments{
\item{mobj}{map_drugs object created with the map_drugs function}

\item{pnr}{pnr}

\item{trend}{whether trend lines should be shown, defaults to FALSE}

\item{scale}{whether scaling should be applied (e.g daily to weekly dose)}
}
\value{
Returns summary plots of dosage trends
}
\description{
Summary function for the map_drugs function
}
