% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablR.R
\name{tablR}
\alias{tablR}
\title{Create frequency tables}
\usage{
tablR(
  data,
  group,
  vars,
  num.vars = NULL,
  test = FALSE,
  total = FALSE,
  numeric = c("medianq1q3", "range"),
  direction = "colwise",
  reference = list(),
  levels = list(),
  labs.groups = list(),
  labs.headings = list(),
  labs.subheadings = list(),
  keep.raw = NULL,
  reverse = F,
  test.stats = c("kwt", "chisq"),
  show.na = NULL,
  censur = F,
  weights,
  digits = 1,
  ama = T,
  style = NULL,
  simplify = list(),
  simplify.remove = c("no", "No", "0"),
  print = F,
  flextable = F
)
}
\arguments{
\item{data}{dataframe}

\item{group}{Grouping variable. If omitted a single column tabel is provided}

\item{vars}{Variables that the grouping variable should be aggregated by}

\item{num.vars}{vector of variables that should be sorted based on numerical order}

\item{test}{Whether statistical tests should be performed (default = FALSE)}

\item{total}{Whether a total column should be included (default = FALSE)}

\item{numeric}{Selection of the type of stats for numerical variables.
Options include: median, q1q3, medianq1q3 (default), iqr, range, mean, sd, meansd, min, max)}

\item{direction}{Direction for percentages (colwise or rowwise)}

\item{reference}{see factR}

\item{labs.groups}{List specifying labels for group variables, see factR}

\item{labs.headings}{List specifying labels for variable names, see factR}

\item{labs.subheadings}{List specifying labels for variable levels. Automatically assigns order, see factR}

\item{keep.raw}{vector of variables that should not be automatically formatted}

\item{reverse}{whether the order of groups should start with the highest level (default = T)}

\item{test.stats}{Vector of length 2 containing statistical tests that should be performed}

\item{show.na}{whether NAs should be displayed ("count") or included in proportions ("pct"), default = NULL}

\item{censur}{whether counts <= 3 should be censored}

\item{weights}{optional name of the column containing weights for weighted summaries}

\item{digits}{number of digits}

\item{ama}{whether percentages >10 should be without digits per AMA journal of style. Default = T.}

\item{style}{character vector specifiying journal stiles. Supported: "lancet".}

\item{simplify}{a list of column names that should be simplified by dropping specified levels (e.g. 0 or "no") for simple output.
Groups of size > 1 should be named such as: list("IHC" = c("cd10", "sox10", "ck"), "necrosis", "margins")}

\item{simplify.remove}{vector of labels that should be removed from the columns assigned in the "simplify" argument. Default = c("no", "0")}

\item{print}{whether the table should be printed in the console}

\item{flextable}{whether the table should be returned as a flextable}
}
\value{
Returns a table as a dataframe or flextable
}
\description{
Wrapper for the tableby function in the Arsenal package
}
\examples{
tablR(population_denmark,
     group = sex,
     vars = c(age_group, population))
}
