% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_match.R
\name{td_match}
\alias{td_match}
\title{td_match}
\usage{
td_match(
  data,
  td_frame,
  index,
  case,
  fu,
  td_date,
  fixed_vars,
  td_vars,
  exclude,
  n_controls = 4,
  seed = 1,
  cores = 4,
  pnr = pnr
)
}
\arguments{
\item{data}{Dataset containing case/control indicator with index date for cases and fixed matching parameters}

\item{td_frame}{Dataset with time dependent matching covariates. Should be multiple rows for each patient with updated values for each parameter.}

\item{index}{Index date for cases, NA for controls}

\item{case}{1/0 indicator for case/control status}

\item{fu}{End of follow up as date}

\item{td_date}{Name of the date column in the time dependent matching covariate dataset}

\item{fixed_vars}{Vector of fixed matching parameters}

\item{td_vars}{Vector of time dependent matching parameters}

\item{exclude}{Vector of parameters that are not allowed to occur before index}

\item{n_controls}{Number of desired controls per case}

\item{seed}{Seed}

\item{cores}{Cores, defaults to 4}

\item{pnr}{Name of PNR column}
}
\value{
Returns the same dataframe as the original but with n_control matches and pairs indicated with a "set" column.
}
\description{
Function for time dependent exact matching
}
