% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizR.R
\name{tickR}
\alias{tickR}
\title{First timestamp for taking time}
\usage{
tickR(cli = T, print = F)
}
\arguments{
\item{cli}{whether the output should be as cli_text (default = T)}

\item{print}{whether the current time should be printet (default = F)}
}
\value{
A timestamp
}
\description{
tickR starts the clock by adding the timestamp "start" to global environment
}
