% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilizR.R
\name{tockR}
\alias{tockR}
\title{Last timestamp for taking time}
\usage{
tockR(format = "diff", start, digits = 2, cli = T)
}
\arguments{
\item{format}{Whether date and time or a time difference should be returned}

\item{digits}{Number of digits on time difference}
}
\value{
Date/time or time difference since tickR()
}
\description{
tockR stops the clock and prints either a date/time or a time difference
}
