% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_td.R
\name{transform_td}
\alias{transform_td}
\title{transform_td}
\usage{
transform_td(data, level = "last")
}
\arguments{
\item{data}{dataframe after matching}

\item{level}{Which value should be picked. last = latest status (default). first = first status. max = maximal status for ranked variables (otherwise the last will be chosen). frequent = most frequent status.}
}
\value{
dataframe with one row per pnr with the last or maximal value of the matched time-dependent variables
}
\description{
Transformation of matching data frame with wide format to long format with one row per pnr
}
