% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatR.R
\name{updatR}
\alias{updatR}
\title{Data reduction, updating and insertion of time-dependent covariates}
\usage{
updatR(td.frame, update.frame, vars, indices = NULL, pnrs = NULL, dt = F)
}
\arguments{
\item{td.frame}{data frame of time-dependent covariates with dates in from/to format}

\item{update.frame}{data.frame of index dates of covariates that should be updated/inserted}

\item{vars}{vector of variables to update/insert}

\item{indices}{vector of index dates for data reduction}

\item{pnrs}{vector of pnrs for data reduction}

\item{dt}{whether the dataframe should be returned as a data.table}
}
\value{
A reduced and updated time-dependent data frame
}
\description{
Data reduction, updating and insertion of time-dependent covariates
}
