% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_search.R
\name{var_search}
\alias{var_search}
\title{var_search}
\usage{
var_search(
  data,
  pattern_list,
  register,
  index_df,
  match = "contains",
  remove = "*_*",
  format = "date",
  interval,
  slice = "first",
  index = index,
  pnr = pnr
)
}
\arguments{
\item{data}{The full register as a data.frame}

\item{pattern_list}{list with list(label = diagnosis code) structure. The pattern can be collected using c() which will be coupled with | as a regex.}

\item{register}{Specify the register}

\item{index_df}{Add index-dates to the patients (optional)}

\item{match}{How specific the diagnosis codes should be used}

\item{remove}{Removal of specific diagnosis codes}

\item{format}{Whether NA/! or dates should be returned}

\item{interval}{Whether the var_search should be in intervals around the index date (e.g. index +/- interval)}

\item{slice}{Which occurence should be extracted}

\item{index}{Name of the index column}

\item{pnr}{Name of the pnr column}
}
\value{
returns a data frame with one or multiple columns based on the labels with either NA/1 or dates if a condition is present
}
\description{
Function to obtain covariates or outcomes from the most common registers
}
