% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewR.R
\name{viewR}
\alias{viewR}
\title{Graphical overview of the structure of a multilevel list}
\usage{
viewR(codelist)
}
\arguments{
\item{codelist}{A codelist of max depth 3}
}
\value{
graphical overview of the levels and groups in the codelist
}
\description{
Graphical overview of the structure of a multilevel list
}
