% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightR.R
\name{weightR}
\alias{weightR}
\title{Weight diagnostics for IPTW}
\usage{
weightR(
  data,
  model,
  treatment,
  vars,
  labs.headings = list(),
  labs.subheadings = list(),
  difference = "standardized",
  weights = "w",
  tail = 50,
  num.vars,
  simplify = F,
  digits = 2
)
}
\arguments{
\item{data}{data.frame}

\item{model}{logistic regression model for estimating propensity scores}

\item{treatment}{treatment variable}

\item{vars}{vector of covariates to estimate weights}

\item{labs.headings}{list of varible names in the format list(new = "old")}

\item{labs.subheadings}{list of variable levels in the format list(var = list(new = "old"))}

\item{difference}{whether the results should be outputtet as absolute or standardized weighted differences}

\item{weights}{name of the column containing weights}

\item{tail}{the number of covariate-groups with the highest weights to show (e.g. top 50)}

\item{num.vars}{pseudo numerical variables for ordering of levels}

\item{simplify}{drop first level of binary variables}
}
\value{
list containing a table of standardized differences (table), overall propensity scores and weights (overall), balance plot (balance) and weight plots (weightst)
}
\description{
Compute weights and diagnostic plots and mean differences
}
