% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cci_time.R
\name{cci_time}
\alias{cci_time}
\title{Time-dependent CCI}
\usage{
cci_time(data, interval = 365.25/2)
}
\arguments{
\item{data}{Full LPR dataset}

\item{interval}{Interval in which the CCIs should be calculated}
}
\value{
cci: Date of first value of each CCI score for each patient
comorb: Date of first presence of each comorbidity
}
\description{
Generation of time-dependent CCI and comorbidities from the Charlson Comorbidity Index
}
