% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closR.R
\name{closR}
\alias{closR}
\title{ClosR}
\usage{
closR(mdose, vec = NULL, str, split = T)
}
\arguments{
\item{mdose}{estimated dose in the patient}

\item{vec}{vector to search for closest value. If NULL, provide str which vec should be created from}

\item{str}{strength of the tablet}

\item{split}{whether the tablet can be split}
}
\value{
Returns the dose closest to the estimated
}
\description{
Pick the closest value from a range in vector.
}
