% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectR.R
\name{collectR}
\alias{collectR}
\title{collectR}
\usage{
collectR(plots, collect = T, nrow = 1, ncol = 3, ...)
}
\arguments{
\item{plots}{List of plots to be collected}

\item{collect}{Whether labels should be collected}

\item{nrow}{Number of rows}

\item{ncol}{Number of colums}

\item{...}{See ggarrange}
}
\value{
Collected plot
}
\description{
Collection of multiple plots into one. Wrapper for ggarrange. Input must be a list of plots.
}
