% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colR.R
\name{colR}
\alias{colR}
\title{Set default color palette}
\usage{
colR(cols = NULL, option = "", show = T, seed = 69)
}
\arguments{
\item{cols}{Vector to set the default color palette. If "default", the default palette is reverted}

\item{option}{miscellaneous features. Includes "default" to reset to default palette, "shuffle" to shuffle the colors}
}
\value{
Adds cancR_palette to the global environment
}
\description{
Set default color palette
}
