% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpR.R
\name{cpR}
\alias{cpR}
\title{cpR}
\usage{
cpR(data, cpr = cpr, extract = F)
}
\arguments{
\item{data}{dataset}

\item{cpr}{cpr-column}

\item{extract}{TRUE if age and date of birth should be extracted}
}
\value{
Returns same dataset with correct CPR numbers and optionally age and date of birth. Invalid CPRs stops the function and returns the invalid CPRs as a vector.
}
\description{
Fixes CPR numbers with removed leading zeros. Works with piping
}
