% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprfix.R
\name{cprfix}
\alias{cprfix}
\title{CPRfix}
\usage{
cprfix(data, cpr)
}
\arguments{
\item{data}{dataset}

\item{cpr}{cpr-column}
}
\value{
Returns same dataset with correct CPR numbers
}
\description{
Fixes CPR numbers with removed leading zeros. Works with piping
}
