% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutR.R
\name{cutR}
\alias{cutR}
\title{Cut numeric variables into intervals}
\usage{
cutR(
  data,
  vars,
  seq.list,
  name.list = list(),
  name.pattern = NULL,
  simplify = T,
  autoformat = T,
  threshold = 1,
  digits = 0,
  dt = F
)
}
\arguments{
\item{data}{dataframe. Works with piping.}

\item{vars}{Vector of vars which should be cut}

\item{seq.list}{List of the split patterns for each var. E.g. list("v1" = seq(0,10), "v2" = round(quantile(df \%>\% pull(age), seq(0,1,0.1)),0)).\
Can be specified using only "median", "quantile", "quartile" etc which will be applied to all vars}

\item{name.list}{list of names for new variables in the format: list("new" = "old"). Names can also be provided as a vector of same length as vars.}

\item{name.pattern}{naming pattern that should automatically be pasted on the end of all the specified variable names in "vars".}

\item{simplify}{whether date intervals should be simplified}

\item{autoformat}{whether > and < should be inserted at the min/max levels and the right limits should be subtracted 0.1/1 (default = T)}

\item{threshold}{number of digits where 0.1 is retracted in autoformat instead of 1 (e.g. threshold = 2 -> 34.9)}

\item{digits}{number of digits for label formatting}

\item{dt}{whether a data.table should be returned (default = F)}
}
\value{
The inputted dataframe with the cut variables
}
\description{
Wrapper for cut() with simplified labels and automatic conversion to factor based on most frequent levels
}
\details{
If a single \code{seq.list} vector or quick name is provided, it is applied
to all \code{vars}.

The pre-specified categorization options are:

\strong{Quantile splits}
\itemize{
\item \code{median}: 50\% (2 groups)
\item \code{tertile}: 33\% (3 groups)
\item \code{quartile}: 25\% (4 groups)
\item \code{pentile}: 20\% (5 groups)
\item \code{decile}: 10\% (10 groups)
\item \code{percentile}: 1\% (100 groups)
}

\strong{Time splits}
\itemize{
\item \code{year}: 1 year
\item \verb{5y}: 5 years
\item \verb{10y}: 10 years
\item \code{half}: 6 months
\item \code{quarter}: 3 months
\item \code{third}: 4 months
}

\strong{BMI}
\itemize{
\item \code{bmi}: splits at 18.5, 25, 30, 35 and 40
}
}
\examples{
redcap_df \%>\%
  datR(c(birth, date_of_surgery)) \%>\%
  cutR(vars = c(age, birth, size, type, date_of_surgery),
       seq.list = list(age = "10y",
                       birth = "year",
                       size = "quartile",
                       type = c(0,2,10),
                       date_of_surgery = "quarter"))
}
