% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutR.R
\name{cutR}
\alias{cutR}
\title{Cut numeric variables into intervals}
\usage{
cutR(data, vars, seqlist, names = list())
}
\arguments{
\item{data}{dataframe. Works with piping.}

\item{vars}{Vector of vars which should be cut}

\item{seqlist}{List of the split patterns for each var. E.g. list("v1" = seq(0,10), "v2" = round(quantile(df \%>\% pull(age), seq(0,1,0.1)),0))}

\item{names}{Whether the split should overwrite or as a new variable}
}
\value{
The inputted dataframe with the cut variables
}
\description{
Wrapper for cut() with simplified labels and automatic conversion to factor based on most frequent levels
}
