% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagR.R
\name{dagR}
\alias{dagR}
\title{Draw directed acyclic graphs (DAGS)}
\usage{
dagR(
  treatment,
  outcome,
  path.list,
  label.fill = "White",
  label.color = "White",
  arrow.fill = "Black",
  arrow.color = "Black",
  arrow.width = 0.7,
  arrow.distance = 0.2,
  curvature = 0,
  label.size = 6,
  dagitty = NULL
)
}
\arguments{
\item{treatment}{name of treatment variable}

\item{outcome}{name of outcome variable}

\item{path.list}{named list of paths to draw in the format list("treatment" = list("to" = "outcome"), "confounder" = list("to" = c("treatment", "outcome")))}

\item{arrow.fill}{color of the arrow heads}

\item{arrow.color}{color of the arrows}

\item{arrow.width}{linewidth of the arrows}

\item{arrow.distance}{percentage reduction in arrow length}

\item{curvature}{curvature of the arrows}

\item{label.size}{text size of the variable nodes}

\item{dagitty}{output from dagitty.net inserted in single quotation marks.}
}
\value{
plot of the specified dag
}
\description{
Function for drawing DAGs. Compatible with dagitty.net via the "dagitty" argument.
}
