% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datR.R
\name{datR}
\alias{datR}
\title{Convert dates from character to date format}
\usage{
datR(data, vars = contains("date"), dt = F)
}
\arguments{
\item{data}{data frame or vector of dates}

\item{vars}{character vector for specifying variables to convert to date format. Default is all columns containing "date"}

\item{dt}{whether a data.table should be returned}
}
\value{
the input data frame with correctly formatted date variables
}
\description{
Convert dates easily without specifying format. The format is identified automatically and converted to standard Year-month-day.
}
\examples{

redcap_df \%>\%
datR(c(birth, date_of_surgery, followup, death_date)) \%>\%
 str

datR(c("2001-02-01", "03-02-2002", 12345))

datR(1234)
}
