% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datR.R
\name{datR}
\alias{datR}
\title{Convert dates from character to date format}
\usage{
datR(data, vars)
}
\arguments{
\item{data}{data frame}

\item{vars}{vector of character vars to convert to date format}
}
\value{
the input data frame with correctly formatted date variables
}
\description{
Convert dates easily without specifying format. The format is identified automatically and converted to standard Year-month-day.
}
