% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodR.R
\name{decodR}
\alias{decodR}
\title{decodR}
\usage{
decodR(codelist, regs = c("pop", "sc", "meta", "dsd"))
}
\arguments{
\item{codelist}{List of lists with registries, diagnosis codes and labels. See example.}
}
\value{
codes: Original codelist without modification \cr
regs: The specified registries for the project. First input in loadR \cr
loadR: The pattern.list in loadR \cr
searchR: The pattern.list in searchR \cr
regex: The codelist converted to regex-expressions with | as separator
}
\description{
Decoding of the main codelist for loading and searching in registries
}
