% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributR.R
\name{distributR}
\alias{distributR}
\title{Assessment of distribution of continuous variables with histograms, QQ-plots and the Shapiro-Wilks test}
\usage{
distributR(data, vars, bins = 1, test = "shapiro")
}
\arguments{
\item{data}{dataframe}

\item{vars}{variables to test. If not specified all numeric variables with more than 5 unique values are assessed}

\item{bins}{binwidth}

\item{test}{whether the Shapiro-Wilks (default) or Kolmogorov-Smirnov test should be performed}
}
\value{
Combined plotframe of histograms, QQ-plots and the Shapiro-Wilks tests
}
\description{
Assessment of distribution of continuous variables with histograms, QQ-plots and the Shapiro-Wilks test
}
