% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatR.multi.R
\name{estimatR.multi}
\alias{estimatR.multi}
\title{estimatR.multi}
\usage{
estimatR.multi(
  data,
  timevar,
  event,
  group,
  names,
  survscale = "AM",
  type = "uni",
  vars = NULL,
  form = NULL,
  time = 120,
  breaks = 12,
  survtime = T,
  proportions = F,
  conditional = F,
  cores = pmin(detectCores(), 4)
)
}
\arguments{
\item{data}{dataset}

\item{names}{Element names of the returned list of models. If missing the "events" names are used.}

\item{survscale}{Whether overall survial should be estimated as survival or all-cause mortality (1-survival)}

\item{type}{Model specification, Can be univariate ("uni"), Age- and sex standardized ("age-sex"), Multivariate with variable selection ("select"). In this case vars should be a vector of the covariates. "custom" allows for free modelling where the form-argument contains the formula.}

\item{vars}{Only applicable when "select" is chosen as type. The variables are added to the model as an additive model}

\item{form}{Only applicable when "custom" is chosen as type. Free specification of the model as the right-hand side of the formula.}

\item{time}{Time horizon of interest. Defaults to 60 (e.g. 5-years)}

\item{breaks}{Interim time points of interest. Defaults to 12 months (1-year gaps)}

\item{survtime}{Whether median time to event should be calculated (default = TRUE)}

\item{cores}{Number of cores for parallel processing}

\item{timevars}{Character vector of time variables. If missing "t_" is assumed to be prefix for all names in the "events" vector}

\item{events}{Character vector of event variables}

\item{groups}{Character vector of grouping variables}
}
\value{
A named list of models with the estimatR function
}
\description{
Convenience function for performing multiple estimatR functions (e.g. multiple outcomes or risk factor). Each argument is recycled to avoid repeated names.
}
