% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatR.multi.R
\name{estimatR.multi}
\alias{estimatR.multi}
\title{Perform multiple estimatR analyses}
\usage{
estimatR.multi(data, timevar, event, group, names, ...)
}
\arguments{
\item{data}{list of dataframe(s)}

\item{timevar}{Character vector of time variables. If missing "t_" is assumed to be prefix for all names in the "events" vector}

\item{event}{Character vector of event variables}

\item{group}{Character vector of grouping variables}

\item{names}{Element names of the returned list of models. If missing the "events" names are used.}

\item{...}{See arguments in estimatR(). Multiple arguments should be inputted as lists (e.g. time = list(60,60,120))}
}
\value{
A named list of models with the estimatR function
}
\description{
Perform multiple estimatR analyses
}
