% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatR.R
\name{estimatR}
\alias{estimatR}
\title{estimatR}
\usage{
estimatR(
  data,
  timevar,
  event,
  group,
  surv,
  multi = FALSE,
  quantiles = T,
  survscale = "OS",
  type,
  vars,
  form,
  horizon = 120,
  timepoints = seq(0, horizon, 6),
  printres = c(seq(0, 60, 12), horizon),
  cores = 4
)
}
\arguments{
\item{data}{Input dataframe}

\item{timevar}{Time-to-event column}

\item{event}{Event indicator. In case of competing risk the structure should be censoring=0, event=1 and death=2. Otherwise
0/1 can be used in case of no competing risks.}

\item{group}{Grouping variable, between which the contrasts are estimated}

\item{surv}{If competing risks = FALSE, otherwise = TRUE.}

\item{multi}{If more than 2 groups are assigned to the "group" argument.}

\item{quantiles}{If median time to event should be returned.}

\item{survscale}{Whether estimates should be presented as Overall survival or All-Cause Mortality}

\item{type}{Assign covariates for adjustment. \cr
uni: No adjustment other than the "group" variable \cr
age-sex: Age and sex standardization \cr
select: Uses "vars" to select variables that should be adjusted for. \cr
custom: The right-hand side of the formula can be customized as a character string.}

\item{vars}{Only applicable if "type" = "select".}

\item{form}{Only applicable if "type" = "custom".}

\item{timepoints}{Defines timepoints of interest at which estimates should be estimated. Defaults to 6 month intervals}

\item{printres}{Defines which time point esimates should be reported.}

\item{cores}{Number of cores for parallel processing. Default = 4.}

\item{t2}{Time-horizon (maximum time point to estimate). Defaults to 120 (10 years).}
}
\value{
Median_surv: Median survival time \cr
Life_table: Event table \cr
Plot_data: Data for plotting CIF curves \cr
HR: Hazard ratio between the groups \cr
Models: Model objects (cause1 and cause2) \cr
diag: Diagnostics for assessing proportionality \cr
Risks: Absolute risk estimates at the specified time points \cr
Risk_differences: Absolute risk differences at the specified time points \cr
Risk_ratios: Absolute risk ratios at the specified time points \cr
Event_proportions: \cr
"Before" estimates the risk of event within a certain timepoint (e.g. x\% of the events occurred within). \cr
"After" estimates the risk of residual events between timepoint x and ten years. \cr
"Window" estimates the percentage of events within six month windows (e.g. x\% of the events occurred between t1 and t2) \cr
Conditional: Conditional risk estimates \cr
Surv: Indicator of the surv-argument
}
\description{
The function uses stratified cause-specific Cox regression and G-computation to estimate the risk of event at a specific
timepoint taking competing risks into account. The function also provides pointwise risk differences and ratios and prepares
the subsequent plotting of the cumulative incidence curves.
The function is a wrapper of the ate-function from riskRegression by T. Gerds et al.
}
