% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatR.R
\name{estimatR}
\alias{estimatR}
\title{estimatR}
\usage{
estimatR(
  data,
  timevar,
  event,
  group,
  quantiles = T,
  survscale = "AM",
  type,
  vars,
  form,
  time = 120,
  breaks = 12,
  cores = pmin(detectCores(), 4)
)
}
\arguments{
\item{data}{dataset}

\item{timevar}{Time variable}

\item{event}{Status indicator}

\item{group}{Grouping variable}

\item{quantiles}{Whether median time to event should be calculated (default = TRUE)}

\item{survscale}{Whether overall survial should be estimated as survival or all-cause mortality (1-survival)}

\item{type}{Model specification, Can be univariate ("uni"), Age- and sex standardized ("age-sex"), Multivariate with variable selection ("select"). In this case vars should be a vector of the covariates. "custom" allows for free modelling where the form-argument contains the formula.}

\item{vars}{Only applicable when "select" is chosen as type. The variables are added to the model as an additive model}

\item{form}{Only applicable when "custom" is chosen as type. Free specification of the model as the right-hand side of the formula.}

\item{time}{Time horizon of interest. Defaults to 60 (e.g. 5-years)}

\item{breaks}{Interim time points of interest. Defaults to 12 months (1-year gaps)}

\item{cores}{Number of cores for parallel processing}
}
\value{
time_to_event: Median survival time \cr
table: Event table \cr
plot_data: Data for plotting CIF curves \cr
hr: Hazard ratio between the groups \cr
models: Model objects (cause1 and cause2) \cr
diag: Diagnostics for assessing proportionality \cr
risks: Absolute risk estimates at the specified time points \cr
differences: Absolute risk difference at the specified time horizons \cr
ratios: Absolute risk ratios at the the specified time horizon \cr
proportions: \cr
"Before" estimates the risk of event within a certain timepoint (e.g. x\% of the events occurred within). \cr
"After" estimates the risk of residual events between timepoint x and ten years. \cr
"Window" estimates the percentage of events within six month windows (e.g. x\% of the events occurred between t1 and t2) \cr
conditional: Conditional risk estimates \cr
info: information on arguments for extraction
}
\description{
Absolute risk estimation and comparison of two or more groups with a time-to-event outcome.
Wrapper for the prodlim and ate-functions in riskRegression.
}
