% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_struct.R
\name{event_struct}
\alias{event_struct}
\title{Event structuring}
\usage{
event_struct(
  data,
  index,
  fu,
  death,
  outcomes,
  namefix = c("{str_remove_all({.col}, 'index_')}",
    "{str_replace_all({.col}, 'index_', 't_')}"),
  unit = "months",
  keep_dates = F
)
}
\arguments{
\item{data}{Dataframe}

\item{index}{Index time point 'T=0' provided in date format}

\item{fu}{End of follow-up or time of death in date format}

\item{death}{Status indicator for death, 0=Alive, 1=Dead}

\item{outcomes}{Vector of single or multiple column names with the event of interest in date format}

\item{namefix}{Vector of length 2. Indicates the name pattern of the event of interest.}

\item{unit}{Whether time-to-event should be reported in months or years.}

\item{keep_dates}{Whether the original event dates should be kept.}
}
\value{
To columns for each event of interest. A status indicator of 0/1/2 and a time-to-event column.
For patients without an event the function returns the time to last follow-up if status = 0 and time to death if status = 2
}
\description{
Converts event dates to status indicator and a time-to-event variable.
The status indicator will be assigned to:
0 if alive and event-free,
1 in the presence of the event of interest
2 in the presence of the competing event, often death
}
