% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractR.R
\name{extractR}
\alias{extractR}
\title{Extraction of key results from the estimatR function}
\usage{
extractR(
  list,
  vars = c("counts", "risks", "diff"),
  format = "long",
  ci = "95\%CI ",
  nsep = " / ",
  sep = " to ",
  total.count = T,
  labs = NULL,
  headings = NULL,
  censur = F,
  risk.digits = 1,
  diff.digits = 1,
  ratio.digits = 1,
  sep.ci = F,
  flextable = F,
  reverse = T
)
}
\arguments{
\item{list}{estimatR object}

\item{vars}{Which outcome measures to provide (counts, risks, differneces and/or ratios)}

\item{format}{Format of the table (long or wide).}

\item{ci}{The format of the confidence intervals (default = 95\%CI)}

\item{nsep}{The separator between event counts (default = /)}

\item{sep}{The separator between all other ranges (default = to)}

\item{censur}{Whether counts <= 3 should be censured (default = FALSE)}

\item{risk.digits}{number of digits on risk estimates}

\item{diff.digits}{number of digits on risk differences}

\item{ratio.digits}{number of digits on risk ratios}
}
\value{
Returns a data frame compatible with flextable
}
\description{
Extraction of key results from the estimatR function
}
