% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractR.R
\name{extractR}
\alias{extractR}
\title{extractR}
\usage{
extractR(
  list,
  vars = c("counts", "risks", "diff"),
  shape = "long",
  ci = "95\%CI ",
  nsep = " / ",
  sep = " to ",
  censur = F
)
}
\arguments{
\item{list}{estimatR object}

\item{vars}{Which outcome measures to provide (counts, risks, differneces and/or ratios)}

\item{shape}{Format of the table (long or wide).}

\item{ci}{The format of the confidence intervals (default = 95\%CI)}

\item{nsep}{The separator between event counts (default = /)}

\item{sep}{The separator between all other ranges (default = to)}

\item{censur}{Whether counts <= 3 should be censured (default = FALSE)}
}
\value{
Returns a data frame compatible with flextable
}
\description{
Automatic extraction of key results from the estimatR function
}
