% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factR.R
\name{factR}
\alias{factR}
\title{factR}
\usage{
factR(
  data,
  vars,
  reference = list(),
  levels = list(),
  labels = list(),
  lab_to_lev = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{vars}{Vector of variables that should be factorized}

\item{reference}{List of variables with the reference level (e.g. list("v1" = "a"))}

\item{levels}{List of variables with the corresponding levels (e.g. list("v1" = "c("a","b","c","d","e")))}

\item{labels}{List of variables with the corresponding labels (e.g. list("v3" = c("e" = "epsilon", "d" = "delta")))}

\item{lab_to_lev}{Whether changing labels should change levels if these are not specified (defaults to TRUE)}
}
\value{
Returns the inputted dataframe with modified factor variables
}
\description{
Convenience function for levelling, labelling and referencing multiple factors in one step. Works with piping.
The order of effects is:
\enumerate{
\item Set as factor (fct_infreq setting levels based on frequency, most common is reference)
\item Set reference levels
\item Set levels manually
\item Set labels
}

The list can be non-complete if no modifications should be passed to the remaining variables
}
