% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/followR.R
\name{followR}
\alias{followR}
\title{Calculate median follow-up time using the inverse Kaplan-Meier method}
\usage{
followR(df, timevar, event, group, time.unit = "m2y")
}
\arguments{
\item{df}{Input dataframe}

\item{timevar}{Time to last follow-up}

\item{event}{Status indicator (censored=0 or dead=1)}

\item{group}{Optional group indicator if follow-up time should be reported separately}
}
\value{
Returns the median follow-up time with IQR for the whole sample or groupwise
}
\description{
Calculate median follow-up time using the inverse Kaplan-Meier method
}
\examples{
followR(analysis_df, time2, event, group = X2)
}
