% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatR.R
\name{formatR}
\alias{formatR}
\title{Auto-formatting of a data frame with layout option for typical levels and labels}
\usage{
formatR(
  data,
  cat_vars = c(case, sex, cci, region, education, income, marital),
  num_vars = c(period, age_group),
  cut_vars = c(age, index),
  labels = list(),
  seqlist = list(),
  names = list(),
  layout = NULL
)
}
\arguments{
\item{data}{dataframe. Works with piping}

\item{cat_vars}{Vector of categorical covariates which should be formatted to factors}

\item{num_vars}{vector of pseudonumerical variables that should be ordered (e.g. 1990-2000, 2000-2010).}

\item{cut_vars}{vector of numerical covariates that should be cut}

\item{labels}{list of labels (e.g. list("var" = c("a" = "apple", "b" = "banana")))}

\item{seqlist}{list of sequences for cutting (e.g. list("age" = seq(0,100,10)))}

\item{names}{names for the cut variable (e.g. list("age" = "age_group")}

\item{layout}{layout preset for common procedures. "Matching" for typical matching studies}
}
\value{
Returns the same dataframe with print-friendly labels for the typical full matching setup.
}
\description{
Auto-formatting of a data frame with layout option for typical levels and labels
}
