% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_stepribbon.R
\name{geom_stepribbon}
\alias{geom_stepribbon}
\title{geom_stepribbon
Copy of the geom_stepribbon function by adibender/pammtools}
\usage{
geom_stepribbon(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  direction = "hv",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
geom_stepribbon
Copy of the geom_stepribbon function by adibender/pammtools
}
