% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidencR.R
\name{incidencR}
\alias{incidencR}
\title{Estimate absolute risks in a single group with competing risks}
\usage{
incidencR(
  data,
  timevar,
  event,
  group,
  survscale = "AM",
  time = 60,
  breaks = 12,
  quantiles = T
)
}
\arguments{
\item{data}{dataframe}

\item{timevar}{time to event column}

\item{event}{event column with either 0/1 structure or 0/1/2 in case of competing risks}

\item{group}{optional grouping column}

\item{survscale}{Whether estimates should be presented as Overall survival or All-Cause Mortality}

\item{time}{time horizon (default=60 months)}

\item{breaks}{Interim time points of interest. Defaults to 12 months (1-year gaps)}

\item{quantiles}{Whether median time to event should be calculated (default = TRUE)}
}
\value{
Life_table: Event table \cr
Plot_data: Data for plotting CIF curves \cr
Median_surv: Median survival time \cr
Surv: Indicator for the type of time-to-event analysis
}
\description{
Estimate absolute risks in a single group with competing risks
}
