% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferencR.R
\name{inferencR}
\alias{inferencR}
\title{Causal inference of time-to-event data}
\usage{
inferencR(
  data,
  treatment,
  timevar,
  event,
  vars,
  outcome.vars,
  censoring.vars,
  method = "tte",
  treat.form = NULL,
  event.form = NULL,
  cens.form = NULL,
  time = 120,
  breaks = 12,
  cause = 1,
  estimator = "AIPTW",
  plot = T,
  survtime = T,
  survscale = "AM",
  bins = 0.5,
  digits = 4,
  event.digits = 2,
  alpha = 0.05,
  weights.digits = 1,
  weights.breaks = 1
)
}
\arguments{
\item{data}{dataframe}

\item{treatment}{column indicating treatment variable. Currently only binary treatments are supported.}

\item{timevar}{Time variable}

\item{event}{Status indicator}

\item{vars}{vector of variables that should be included in the treatment, censoring and outcome models}

\item{outcome.vars}{vector of variables that should only be included in the outcome model}

\item{censoring.vars}{vector of variables that should only be included in the censoring model}

\item{method}{whether the outcome is "tte" (time-to-event) or "binary" (default = "tte")}

\item{treat.form}{custom formula for the RHS of the treatment model}

\item{event.form}{custom formula for the RHS of the outcome model}

\item{cens.form}{custom formula for the RHS of the censoring model}

\item{time}{Time horizon of interest. Defaults to 60 (e.g. 5-years)}

\item{breaks}{Interim time points of interest. Defaults to 12 months (1-year gaps)}

\item{cause}{cause of interest, default = 1}

\item{estimator}{whether the estimator should be "IPTW", "GFORMULA" or "AIPTW" (default)}

\item{plot}{whether plot data should be estimated for time optimization (default = T)}

\item{survtime}{Whether median time to event should be calculated (default = F)}

\item{survscale}{Whether overall survial should be estimated as survival or all-cause mortality (1-survival)}

\item{bins}{bins for the weights plot, default = 0.5}

\item{digits}{for rounding of eventtimes}

\item{event.digits}{whether eventtimes should be rounded. Default is 2 to preserve exact times}

\item{alpha}{alpha level for the estimation of confidence intervals and p-values. Default = 0.05}

\item{weights.digits}{the number of digits for categorization of weights}

\item{weights.breaks}{breaks for categorization of weights}
}
\value{
table: Event table \cr
time_to_event: Median survival time \cr
plot_data: Data for plotting CIF curves \cr
risks: Absolute risk estimates at the specified time points \cr
differences: Absolute risk difference at the specified time horizons \cr
ratios: Absolute risk ratios at the the specified time horizon \cr
counts: Event and group counts in the contrasted groups \cr
weights: Object containing: \cr
data: the raw dataset with propensity scores (ps), weights (w) and categorized weighting groups (wgroup) \cr
table_iptw: a table of the crude and iptw adjusted covariates \cr
table_strat: a table of the weigthing groups stratified on treatment groups \cr
plot: plot of the propensity scores and weights for each group
models: Model objects (cause1 and cause2) \cr
diagnostics: Diagnostics for assessing proportionality \cr
info: information on arguments for extraction
}
\description{
Average treatment effect of point intervention with either IPTW, GFORMULA or AIPTW. Wrapper for the ATE function in riskRegression.
}
