% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteratR.R
\name{iteratR}
\alias{iteratR}
\title{Perform multiple estimatR analyses}
\usage{
iteratR(data, names, method, ...)
}
\arguments{
\item{data}{list of dataframe(s)}

\item{names}{vector of names for the returned list of models. If missing the "events" names are used.}

\item{method}{which cancR function to use. Choose between "estimatR", "incidencR", "clustR" and "inferencR".}

\item{...}{See arguments in the specific function documentations. Multiple arguments should be inputted as lists (e.g. time = list(60,60,120))}
}
\value{
A named list of models of the type specified in "method"
}
\description{
Perform multiple estimatR analyses
}
