% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinR.R
\name{joinR}
\alias{joinR}
\title{Join two or more dataframes}
\usage{
joinR(..., by, type = "left", dt = F)
}
\arguments{
\item{...}{dataframes or list of dataframes to be joined}

\item{by}{variable to join on}

\item{type}{type of join (e.g. left (default), right, full and inner)}

\item{dt}{whether the returned dataframe should be a data.table}
}
\value{
a single joined dataframe
}
\description{
Join two or more dataframes as a left, right, full or inner join. The dataframes can be provided separately or as a list.
}
