% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listR.R
\name{listR}
\alias{listR}
\title{Routine modifications of lists}
\usage{
listR(input, type, layer, collapse = F, ...)
}
\arguments{
\item{input}{the list or vector}

\item{type}{type of modification of the list. See details}

\item{layer}{vector of integers indicating which layer to remove if type is "peel" or to keep if type is "pick". Layer 1 is the top layer}

\item{collapse}{whether duplicated elements should be collapsed (defualt = F)}

\item{...}{arguments for the rrapply function}
}
\value{
returns a modified list based on the "type" argument
}
\description{
Routine modification of lists such as reverse values and names for use in other functions.
The functions is also a wrapper for the rrapply and map_depth functions with the dots argument.
}
\details{
"reverse" reverses the values and names of a list which is used as inputs in functions such as str_replace.
"vec2list" converts a vector to a list with names corresponding to the vector elements
"peel" and "pick" depends on "layer" and either drops or keeps the specified vector.
}
\examples{

reverse_list <- list("first" = "a1",
                     "second" = "b2")

listR(reverse_list, type = "reverse")
}
