% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listR.R
\name{listR}
\alias{listR}
\title{Routine modifications of lists}
\usage{
listR(list, type, ...)
}
\arguments{
\item{list}{the list to modify}

\item{type}{type of modification of the list. See details}

\item{...}{arguments for the rrapply function}
}
\value{
returns a modified list based on the "type" argument
}
\description{
Routine modification of lists such as reverse values and names for use in other functions.
The functions is also a wrapper for the rrapply and map_depth functions with the dots argument.
}
\details{
"reverse" reverses the values and names of a list which is used as inputs in functions such as str_replace.
}
