% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadR.R
\name{loadR}
\alias{loadR}
\title{Load registers}
\usage{
loadR(
  regs,
  pattern.list = NULL,
  pattern.list2 = NULL,
  n = NULL,
  id.filter = NULL,
  keep.list = NULL,
  vars.list = NULL,
  lmdb.start = 1995,
  lmdb.stop = 2023,
  simulation = F,
  cores = 4,
  ...
)
}
\arguments{
\item{regs}{which registers should be loaded. Default is all (lpr, pop, pato, cancer, lmdb and opr)}

\item{pattern.list}{list of vectors of diagnoses codes for each register in the format ("lpr" = c("DC92", "DC21"))}

\item{pattern.list2}{supplemental pattern if multiple columns should be filtered. Works the same way as pattern}

\item{n}{number of observations that should be loaded}

\item{id.filter}{optional possibility to limit the registers to a defined patient population of PNRs}

\item{keep.list}{which variables should be kept provided as a list("lpr" = c("vars")).}

\item{vars.list}{which columns should the pattern filter be applied to. Defaults to diag, atc, opr and c_morfo3,}

\item{lmdb.start}{first year of LMDB}

\item{lmdb.stop}{last year of LMDB}

\item{simulation}{whether the registers should be simulated}

\item{...}{arguments passed to simulatR()}
}
\value{
Returns the selected registers to the global environment
}
\description{
Simple loading function of the most used registers including variable selection, regex filtering and id_list filtering.
}
