% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missR.R
\name{missR}
\alias{missR}
\title{Overview of NAs in a dataframe}
\usage{
missR(data, vars, id = id, print = T)
}
\arguments{
\item{data}{data frame}

\item{vars}{vars where the na check should be beformed. If missing the whole data frame is analysed}

\item{id}{name of the id-column. If missing a guess at the most common names is attempted.}

\item{print}{whether the NA check should be printed in the console}
}
\value{
Prints whether any NAs are detected and returns a data frame with IDs and columns with NA
}
\description{
Overview of NAs in a dataframe
}
