% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode}
\usage{
mode(x, ties = "first", na.rm = T)
}
\arguments{
\item{x}{Vector of values}

\item{ties}{If two or more values are equally common, which should be chosen. Default is first.}

\item{na.rm}{Whether NAs should be removed, defaults to TRUE}
}
\value{
The most common value of the vector excluding NAs
}
\description{
Get the mode (most common value) of a vector.
}
