% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbR.R
\name{numbR}
\alias{numbR}
\title{Format numeric vectors}
\usage{
numbR(numbers, digits = 1, nsmall)
}
\arguments{
\item{numbers}{numeric value or vector for formatting}

\item{digits}{number of digits}

\item{nsmall}{number of zero-digits}
}
\value{
returns a vector of same lentgh with formatted digits
}
\description{
Format numeric vectors
}
