% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotR.R
\name{plotR}
\alias{plotR}
\title{Autoplot for estimatR, inferencR and clustR}
\usage{
plotR(
  list,
  y = 100,
  col = cancR_palette,
  table.col = "#616161",
  risk.col = F,
  time.unit = "m2y",
  labs = levels,
  print.est = TRUE,
  contrast = "rd",
  se = T,
  p.values = T,
  time.to.event = "none",
  cens.lines = F,
  style = NULL,
  linewidth = 0.8,
  title = "",
  title.size = 7,
  title.shift = c(0, 0),
  x.title = unit,
  x.title.size = 6,
  x.title.shift = 0,
  x.text.size = 6,
  y.title = "Risk of Event (\%)",
  y.title.size = 6,
  y.title.shift = 0,
  y.text.size = 6,
  res.size = 5,
  res.shift = c(0, 0),
  res.spacing = 1,
  res.digits = 1,
  box = T,
  box.shift = 0,
  box.fill = "White",
  box.color = "Black",
  box.linewidth = 0.8,
  contrast.digits = 1,
  table = c("event", "risk"),
  table.space = 1,
  table.padding = 1,
  table.title.size = 6,
  table.text.size = 5,
  table.linewidth = 0.8,
  legend.pos = c(0.5, 0.9),
  legend.size = 16,
  tscale = 1,
  censur = F
)
}
\arguments{
\item{list}{an object of class estimatR, inferencR or clustR}

\item{y}{Upper limit for y-axis}

\item{col}{Vector of colors}

\item{table.col}{Grid color}

\item{risk.col}{Whether risk table numbers should be colored (T/F)}

\item{time.unit}{Specification of the time-unit and optional conversion. Conversions include Months to years ("m2y"), days to years ("d2y") and days to months ("d2m")}

\item{labs}{Character vector of similar length to the number of levels in the group with labels. Reference is first.}

\item{print.est}{Whether absolute risks at the time horizon should be printet. Defaults to TRUE}

\item{contrast}{The type of contrast that should be provided. Includes risk difference ("rd", default), risk ratio ("rr"), hazard ratio ("hr") or "none".}

\item{se}{whether the confidence interval should be shown}

\item{p.values}{whether p-values should be printed in the results, default = T}

\item{time.to.event}{type of time to event line, choose between "vertical", "horizontal", "both" or "none" (default)}

\item{cens.lines}{whether censoring points should be printed}

\item{style}{the formatting style of the contrast. Currently JAMA and italic}

\item{linewidth}{thickness of the risk curve lines}

\item{title}{Plot title}

\item{title.size}{Plot title size}

\item{title.shift}{vector of XY shifting of the plot title}

\item{x.title}{X-axis title}

\item{x.title.size}{X-axis title size}

\item{x.title.shift}{X-axis vertical shift}

\item{x.text.size}{X-axis text size}

\item{y.title}{Y-axis title}

\item{y.title.size}{Y-axis title size}

\item{y.title.shift}{Y-axis title horizontal shift}

\item{y.text.size}{Y-axis text.size}

\item{res.size}{Size of the results}

\item{res.shift}{Vector of XY shifting of the results}

\item{res.spacing}{Vertical spacing between results}

\item{res.digits}{Number of digits on the risk estimates}

\item{box}{whether there should be a box around the results}

\item{box.shift}{Horizontal shifting of the right end of the box}

\item{box.fill}{fill color for the box}

\item{box.color}{border color for the box}

\item{box.linewidth}{Results box linewidth}

\item{contrast.digits}{Number of digits on the contrasts}

\item{table}{Which parts of the risk table should be provided ("event", "risk", "none"). Default is c("event", "risk")}

\item{table.space}{Spacing between counts in risk table}

\item{table.padding}{Spacing between lines and first/last rows in the risk table}

\item{table.title.size}{Risk table titles size}

\item{table.text.size}{Risk table text size}

\item{table.linewidth}{Risk table linewidth}

\item{legend.pos}{XY vector of legend position in percentage}

\item{tscale}{Global size scaler}

\item{censur}{Whether values <= 3 should be censored. Default = FALSE}
}
\value{
Plot of the adjusted cumulative incidence or Kaplan-Meier curve
}
\description{
Autoplot for estimatR, inferencR and clustR
}
\examples{
#Risk in one group

t1 <- estimatR(analysis_df,
timevar = ttt,
event = event)

plotR(t1)

#Risks in multiple groups
t2 <- estimatR(analysis_df,
timevar = ttt,
event = event,
group = X2)

plotR(t2)


}
