% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotR.R
\name{plotR}
\alias{plotR}
\title{plotR}
\usage{
plotR(
  list,
  y = 100,
  col = c("#9B62B8", "#224B87", "#67A8DC", "#D66ACE"),
  table.col = "#616161",
  time_unit = "m2y",
  labgrp = levels,
  print.est = TRUE,
  contrast = "rd",
  title = "",
  title.size = 7,
  x.title = unit,
  x.title.size = 6,
  x.text.size = 6,
  y.title = "Risk of Event (\%)",
  y.title.size = 6,
  y.title.shift = 0,
  y.text.size = 6,
  res.size = 5,
  res.shift = c(0, 0),
  res.spacing = 1,
  res.digits = 1,
  contrast.digits = 1,
  table = c("event", "risk"),
  table.space = 1,
  table.padding = 1,
  table.title.size = 6,
  table.text.size = 5,
  table.linewidth = 1,
  legend.pos = c(0.5, 0.95),
  tscale = 1,
  censur = F
)
}
\arguments{
\item{list}{an object of class estimatR, incidencR or clustR}

\item{y}{Upper limit for y-axis}

\item{col}{Vector of colors}

\item{table.col}{Grid color}

\item{time_unit}{Specification of the time-unit and optional conversion. Conversions include Months to years ("m2y"), days to years ("d2y") and days to months ("d2m")}

\item{print.est}{Whether absolute risks at the time horizon should be printet. Defaults to TRUE}

\item{contrast}{The type of contrast that should be provided. Includes risk difference ("rd", default), risk ratio ("rr"), hazard ratio ("hr") or "none".}

\item{title}{Plot title}

\item{title.size}{Plot title size}

\item{x.title}{X-axis title}

\item{x.title.size}{X-axis title size}

\item{x.text.size}{X-axis text size}

\item{y.title}{Y-axis title}

\item{y.title.size}{Y-axis title size}

\item{y.title.shift}{Y-axis title horizontal shift}

\item{y.text.size}{Y-axis text.size}

\item{res.size}{Size of the results}

\item{res.shift}{Vector of XY shifting of the results}

\item{res.spacing}{Vertical spacing between results}

\item{res.digits}{Number of digits on the risk estimates}

\item{contrast.digits}{Number of digits on the contrasts}

\item{table}{Which parts of the risk table should be provided ("event", "risk", "none"). Default is c("event", "risk")}

\item{table.space}{Spacing between counts in risk table}

\item{table.padding}{Spacing between lines and first/last rows in the risk table}

\item{table.title.size}{Risk table titles size}

\item{table.text.size}{Risk table text size}

\item{table.linewidth}{Risk table linewidth}

\item{legend.pos}{XY vector of legend position in percentage}

\item{tscale}{Global size scaler}

\item{censur}{Whether values <= 3 should be censored. Default = FALSE}

\item{labs}{Character vector of similar length to the number of levels in the group with labels. Reference is first.}
}
\value{
Plot of the adjusted cumulative incidence or Kaplan-Meier curve
}
\description{
Automatic plot function for the functions estimatR, incidencR and clustR
}
