% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powR.R
\name{powR}
\alias{powR}
\title{Converter from Rscript to powershell format}
\usage{
powR(text)
}
\arguments{
\item{text}{R script that needs to be converted to powershell format}
}
\value{
Printed text compatible with powershell
}
\description{
Converter from Rscript to powershell format
}
