% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvertR.R
\name{pvertR}
\alias{pvertR}
\title{Format p-values to AMA manual of style}
\usage{
pvertR(x)
}
\arguments{
\item{x}{A p-value}
}
\value{
Prints the raw p-value according to AMA manual of style
}
\description{
Format p-values to AMA manual of style
}
