% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvertR.R
\name{pvertR}
\alias{pvertR}
\title{Format p-values to AMA manual of style}
\usage{
pvertR(x, na = "NA")
}
\arguments{
\item{x}{A p-value}

\item{na}{the print of NA values, default = "NA.}
}
\value{
Prints the raw p-value according to AMA manual of style
}
\description{
Format p-values to AMA manual of style
}
