% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvertR.R
\name{pvertR}
\alias{pvertR}
\title{pvertR}
\usage{
pvertR(x)
}
\arguments{
\item{x}{A p-value}
}
\value{
Prints the raw p-value according to AMA manual of style
}
\description{
The function takes a raw p-value and prints it according to the AMA manual of style
}
