% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizR.R
\name{randomizR}
\alias{randomizR}
\title{Generate stratified, block randomized allocation sequence}
\usage{
randomizR(
  strata,
  block.size,
  block.probabilities = NULL,
  replications = 50,
  allocation.levels = c("1", "2"),
  allocation.name = "allocation",
  block.delay = 10,
  token,
  seed = 1,
  print = F
)
}
\arguments{
\item{strata}{named list of strata levels (e.g. list("age" = c(1,2), "sex" = c("f","m"))). If missing, an API token can be provided (see "token")}

\item{block.size}{vector of block size(s).}

\item{block.probabilities}{vector of block size probabilities}

\item{replications}{length of the allocation sequence in each stratum}

\item{allocation.levels}{vector of the allocated treatment levels, default = c("1", "2")}

\item{allocation.name}{column name for the allocated treatment levels, default = "allocation"}

\item{block.delay}{number of blocks that is fixed to the first block.size to avoid large blocks in the beginning of the sequence}

\item{token}{token for redcap API to automatically extract strata levels from redcap. If strata is a vector of variables, token must be provided.}

\item{seed}{seed for reproducibility}

\item{print}{whether a full report for each stratum should be printed}
}
\value{
a data frame containing rows equal to the product of unique strata and replications.
}
\description{
Stratified block randomization sequence with optional redcap API. Each stratum will have its own unique allocation sequence.
}
