% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readR.R
\name{readR}
\alias{readR}
\title{Load csv, excel and rds files}
\usage{
readR(path)
}
\arguments{
\item{path}{path for the file to load}
}
\value{
the given path imported as a data frame
}
\description{
Wrapper for the fread, readxl and readRDS functions with automatic detection of file extension
}
