% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodR.R
\name{recodR}
\alias{recodR}
\title{Recode multiple variables}
\usage{
recodR(data, namelist, match = "exact", replace = F)
}
\arguments{
\item{data}{dataframe}

\item{namelist}{list og variables with named vector(s) containing diagnosis codes (e.g. list(var1 = list(name = diagnosis codes)))}

\item{match}{Whether the provided diagnosis codes should be matched exactly, start with/end with, be bounded by or contain (default)}

\item{replace}{whether multiple matches should be replaced directly (such as 1,4 to head, arm)}
}
\value{
the input data frame with recoded variables
}
\description{
Recode multiple variables
}
\examples{
set.seed(1)
df <-
data.frame(diag = sample(c("DX123", "DC123", "DC234", "DG123", "DG234"), 20, replace=TRUE),
           type = sample(c("DY", "DY234", "DY123"), 20, replace=TRUE),
           type2 = sample(c("DC123", "DC234", "DG123"), 20, replace=TRUE),
           split = c("1,11", "2,10", "2,4", "2,15"))

            df \%>\%
  recodR(list("diag" = list("KOL" = "DX123",
                            "Astma" = c("DC123", "DC2"),
                            "AMI" = c("DG123", "DG234"))),
         match = "exact")
df \%>\%
  factR(c(diag, type, type2)) \%>\%
  recodR(list("diag" = list("KOL" = "DX123",
                            "Astma" = c("DC123", "DC234"),
                            "AMI" = list("DG123", "DG234"))),
         match = "exact")

df \%>\%
  recodR(list("split" = list("one" = "1",
                             "four" = "4",
                             "ten" = "10",
                             "fifteen" = "15")),
         match = "boundary")

df \%>\%
  recodR(list("split" = list("one" = "1",
                             "two" = "2",
                            "eleven" = "11",
                             "four" = "4",
                             "ten" = "10",
                             "fifteen" = "15")),
         match = "boundary",
         replace=T)

}
