% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapR.R
\name{redcapR}
\alias{redcapR}
\title{Autoformatting of redcap exports}
\usage{
redcapR(data, dictionary, namelist = list(), autoformat = T, formatlist = NULL)
}
\arguments{
\item{data}{raw redcap dataset}

\item{dictionary}{data dictionary in semicolon-separated .csv format}

\item{namelist}{optional list for manual labelling e.g. list("observer" = list("John" = "1", "Me" = "2"))}

\item{autoformat}{whether all labels should be in lowercase and underscores as spaces}

\item{formatlist}{optional list for recoding common values such as positiv -> 1/pos/yes e.g. list("pos" = "positive", "no" = "Not Present")}
}
\value{
relabelled redcap dataset with correctly formatted dates
}
\description{
Automatic grabbing of labels from data dictionary and conversion to correct date format for exported redcap datasets.
The data dictionary can be downloaded under: Project Home and Design -> Dictionary -> Download the Data Dictionary
}
