% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportR.R
\name{reportR}
\alias{reportR}
\title{reportR}
\usage{
reportR(
  data,
  casename,
  vars = c(period, age_group, sex, education, income, cci, region, marital),
  table = F,
  plot = F,
  type = "simple",
  cols = cancR_palette,
  headings = list(),
  layout = "vertical",
  vjust = -0.5,
  text.color = "White",
  ...
)
}
\arguments{
\item{data}{matched dataset}

\item{casename}{name or number indicating cases (e.g. "1" or "CLL")}

\item{vars}{vars that should be in the table}

\item{table}{whether a table should be made (default = F)}

\item{plot}{whether a plot should be made (default = F)}

\item{type}{whether non-matched counts should be collapsed to "unmatched" ("simple") or remain stratified ("full")}

\item{cols}{color palette (default is cancR_palette)}

\item{headings}{List specifying labels for variable names}

\item{layout}{layout of the bar chart (horizontal or vertical (default))}

\item{vjust}{vertical adjustment of the counts (pct) labels}

\item{text.color}{label colors}

\item{...}{passed to tablR}
}
\value{
Prints the matching report, table and plot. Returns af list of the table and plot.
}
\description{
Automatic tablR function for overview of matched and unmatched cases
}
