% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportR.R
\name{reportR}
\alias{reportR}
\title{reportR}
\usage{
reportR(
  data,
  casename,
  vars = c(period, age_group, sex, education, income, cci, region, marital),
  table = F,
  plots = F,
  ...
)
}
\arguments{
\item{data}{matched dataset}

\item{casename}{name or number of cases (e.g. 1 or "CLL")}

\item{vars}{vars that should be in the table}

\item{n.controls}{n.controls from the matchR function}

\item{labels}{List specifying labels of the specific labels for each variable}

\item{headings}{List specifying labels for variable names}

\item{filename}{}
}
\value{
Returns a table of the matching overview and exports a word-file (optional if filename is provided)
}
\description{
Automatic tablR function for overview of matched and unmatched cases
}
