% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollR.R
\name{rollR}
\alias{rollR}
\title{Assign rolling ID.}
\usage{
rollR(data, vars = id, label = order)
}
\arguments{
\item{data}{dataset}

\item{vars}{column(s) which unique ids should be based on}

\item{label}{label for new unique id column}
}
\value{
adds a new column to the dataset with unique ids based on original id conditional on a sorting
}
\description{
Useful after sorting a dataset, which changes the ID order. Works with piping
}
