% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowR.R
\name{rowR}
\alias{rowR}
\title{Perform rowwise operations}
\usage{
rowR(
  data,
  vars,
  type,
  label,
  collapse = "",
  na.rm = T,
  filter = NULL,
  direction,
  drop = F,
  dt = F
)
}
\arguments{
\item{data}{dataset}

\item{vars}{vector of vars the rowwise operation should be perfomed. If missing, all columns are assigned}

\item{type}{pmin, pmax, all.na, any.na, sum.na, sum or fill}

\item{label}{label of the new variable. If missing, type is inserted (e.g. pmin)}

\item{na.rm}{whether NAs should be removed (default = T)}

\item{filter}{whether rows with all.na or any.na should be removed}

\item{direction}{the direction of the rowwise fill. Can be "left", "rigth", "leftright" and "rightleft". Corresponds to "updown".}

\item{drop}{whether vars should be dropped after the rowwise operation}

\item{dt}{whether the dataframe should be returned as data.table}
}
\value{
the parent data frame with a new variable based on the selected rowwise operator or a filtered dataset
}
\description{
Perform rowwise operations
}
