% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowR.R
\name{rowR}
\alias{rowR}
\title{Perform rowwise operations}
\usage{
rowR(
  data,
  vars,
  type,
  new,
  na.rm = T,
  filter = NULL,
  match = "contains",
  direction
)
}
\arguments{
\item{data}{dataset}

\item{vars}{vector of vars the rowwise operation should be perfomed}

\item{type}{pmin, pmax, all.na, any.na, sum.na, sum}

\item{new}{name of the new variable. If missing, type is inserted (e.g. pmin)}

\item{na.rm}{whether NAs should be removed (default = T)}

\item{filter}{whether rows with all.na or any.na should be removed}

\item{match}{whether the provided diagnosis codes should be matched exactly, start with/end with or contain (default)}

\item{direction}{the direction of the rowwise fill. Can be "left", "rigth", "leftright" and "rightleft". Corresponds to "updown".}
}
\value{
the parent data frame with a new variable based on the selected rowwise operator or a filtered dataset
}
\description{
Perform rowwise operations
}
