% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowR.R
\name{rowR}
\alias{rowR}
\title{rowR}
\usage{
rowR(data, vars, type, new, na.rm = T, filter = F, match = "contains")
}
\arguments{
\item{data}{dataset}

\item{vars}{vector of vars the rowwise operation should be perfomed}

\item{type}{pmin, pmax, all.na, any.na, sum.na, sum}

\item{new}{name of the new variable. If missing, type is inserted (e.g. pmin)}

\item{na.rm}{whether NAs should be removed (default = T)}

\item{filter}{whether rows with all.na or any.na should be removed}
}
\value{
the parent data frame with a new variable based on the selected rowwise operator or a filtered dataset
}
\description{
Common rowwise operations
}
