% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savR.R
\name{savR}
\alias{savR}
\title{SavR}
\usage{
savR(
  plot,
  name,
  width = 154,
  height = 60,
  unit = "mm",
  scale = 2,
  dpi = 1200,
  device = NULL,
  compression = "lzw",
  formats = c("pdf", "svg", "tiff", "jpg")
)
}
\arguments{
\item{plot}{Plot to plot}

\item{name}{File name of saved plot without extension}

\item{width}{width}

\item{height}{heigth}

\item{unit}{mm or cm}

\item{scale}{to fit}

\item{dpi}{resolution}

\item{device}{cairo for special occasions}

\item{compression}{for tiff}

\item{formats}{choose between pdf, svg, tiff, jpg and png}
}
\value{
Saves plot automatically in current project folder
}
\description{
Wrapper for ggsave with default settings and automatic saving of multiple formats
}
