% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisR.R
\name{summarisR}
\alias{summarisR}
\title{summarisR}
\usage{
summarisR(data)
}
\arguments{
\item{data}{Dataset}
}
\value{
Returns bar charts or density plots depending on format. Numerical variables with less than 5 unique values are considered as factors.
}
\description{
Graphical overview of entire dataset
}
