% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimmR.R
\name{swimmR}
\alias{swimmR}
\title{Plot individual patient trajectories from time-to-event data (Swimmer plot)}
\usage{
swimmR(
  data,
  event,
  strata,
  strata.labels = list(),
  strata.levels = list(),
  horizon = 60,
  breaks = 6,
  table.list = NULL,
  table.cols = c("gainsboro", "grey", "#BCCFE8"),
  cols = list(),
  shapes = list(),
  labs,
  id = "id"
)
}
\arguments{
\item{data}{data frame containg status and time-to-event columns}

\item{event}{columns containing status indicators. Needs to have corresponding "t_" prefix}

\item{strata}{optional stratification by a group}

\item{strata.labels}{named list of labels for the strata in the format list(new = "old)}

\item{strata.levels}{character vector of levels for the strata}

\item{horizon}{time horizon}

\item{breaks}{breaks between time zero and horizon}

\item{table.list}{named list for each stratum with a named list of rows and sequence of x-values}

\item{table.cols}{vector of the colors of the tables}

\item{cols}{named list with colors for the symbols}

\item{shapes}{named list with shapes}

\item{labs}{named list with labels for the events}

\item{id}{name of the id variable}
}
\value{
a swimmer plot
}
\description{
Plot individual patient trajectories from time-to-event data (Swimmer plot)
}
