% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablR.R
\name{tablR}
\alias{tablR}
\title{tablR}
\usage{
tablR(
  data,
  group,
  vars,
  test = FALSE,
  total = FALSE,
  numeric = c("median", "q1q3", "range"),
  direction = "colwise",
  levels = list(),
  labels = list(),
  reference = list(),
  headings = NULL,
  test_stats = c("kwt", "chisq"),
  show_na = FALSE,
  filename = NULL,
  sort.numeric = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{group}{Grouping variable. If omitted a single column tabel is provided}

\item{vars}{Variables that the grouping variable should be aggregated by}

\item{test}{Whether statistical tests should be performed (default = FALSE)}

\item{total}{Whether a total column should be included (default = FALSE)}

\item{numeric}{Selection of the type of stats for numerical variables (e.g. median, q1q3, range, mean, sd)}

\item{direction}{Direction for percentages (colwise or rowwise)}

\item{labels}{List specifying labels of the specific labels for each variable}

\item{headings}{List specifying labels for variable names}

\item{test_stats}{Vector of length 2 containing statistical tests that should be performed}

\item{show_na}{Whether NAs should be presented}

\item{filename}{The filname of the exported word-table}
}
\value{
Returns a table and exports a word-file (optional if filename is provided)
}
\description{
Wrapper for the tableby function in the Arsenal package
}
