% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablR.R
\name{tablR}
\alias{tablR}
\title{Create baseline table}
\usage{
tablR(
  data,
  group,
  vars,
  num.vars,
  test = FALSE,
  total = FALSE,
  numeric = c("median", "q1q3", "range"),
  direction = "colwise",
  reference = list(),
  levels = list(),
  labs.groups = list(),
  labs.headings = list(),
  labs.subheadings = list(),
  reverse = F,
  test.stats = c("kwt", "chisq"),
  show.na = FALSE,
  censur = F,
  digits = 1,
  simplify = list(),
  print = F,
  flextable = F
)
}
\arguments{
\item{data}{dataframe}

\item{group}{Grouping variable. If omitted a single column tabel is provided}

\item{vars}{Variables that the grouping variable should be aggregated by}

\item{num.vars}{vector of variables that should be sorted based on numerical order}

\item{test}{Whether statistical tests should be performed (default = FALSE)}

\item{total}{Whether a total column should be included (default = FALSE)}

\item{numeric}{Selection of the type of stats for numerical variables (e.g. median, q1q3, range, mean, sd)}

\item{direction}{Direction for percentages (colwise or rowwise)}

\item{reference}{List specifying reference group for each variable}

\item{labs.headings}{List specifying labels for variable names}

\item{reverse}{whether the order of groups should start with the highest level (default = T)}

\item{test.stats}{Vector of length 2 containing statistical tests that should be performed}

\item{show.na}{Whether NAs should be presented}

\item{censur}{whether counts <= 3 should be censored}

\item{digits}{number of digits}

\item{labels}{List specifying labels of the specific labels for each variable}
}
\value{
Returns a table as a dataframe or flextable
}
\description{
Wrapper for the tableby function in the Arsenal package
}
\examples{
tablR(population_denmark,
     group = sex,
     vars = c(age_group, population))
}
