% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_td.R
\name{transform_td}
\alias{transform_td}
\title{transform_td}
\usage{
transform_td(data, level = "last")
}
\arguments{
\item{data}{dataframe after matching}

\item{level}{whether the last value or maximal value should be picked. Maximal applies for ranked variables. Otherwise the last will be chosen.}
}
\value{
dataframe with one row per pnr with the last or maximal value of the matched time-dependent variables
}
\description{
Transformation of matching data frame with wide format to long format with one row per pnr
}
