% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumR.R
\name{tumR}
\alias{tumR}
\title{Map unique tumors}
\usage{
tumR(
  data,
  tumor,
  loc.exact = F,
  pnr = pnr,
  date = date,
  dt = F,
  verbose = F,
  tumor_distance = c(-90, 365.25 * 5),
  meta_distance = 365.25 * 2
)
}
\arguments{
\item{data}{dataframe of data from the Danish National Pathology Register}

\item{tumor}{named list of tumors with vectors of snomed codes (e.g. list("pcc" = c("m0703", "m805"), "bcc" = "m809"))}

\item{loc.exact}{whether tumors should match on exact location to be considered linked}

\item{dt}{whether a data.table should be returned (default = F)}

\item{verbose}{whether detailed status for each patient should be printed (for debugging)}

\item{tumor_distance}{the interval where a metastasis is considered relevant}

\item{meta_distance}{the distance between metastases that classifies a cluster}

\item{id}{name of the main identifier, default = pnr}
}
\value{
a data.frame of unique tumors with date, diagnosis and location
}
\description{
Customized function for the Danish Pathology Register to map unique tumors and outcomes taking reexcisions and changes in diagnosis into account.
}
