% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumR.R
\name{tumR}
\alias{tumR}
\title{Map unique tumors}
\usage{
tumR(data, tumor, loc.exact = F, cores = NULL, dt = F)
}
\arguments{
\item{data}{dataframe of data from the Danish National Pathology Register}

\item{tumor}{character vector containing prefixes for tumors of interest (e.g. m807)}

\item{loc.exact}{whether tumors should match on exact location to be considered linked}

\item{cores}{number of cores for parallel processing}

\item{dt}{whether a data.table should be returned (default = F)}
}
\value{
a data.frame of unique tumors with date, diagnosis and location
}
\description{
Customized function for the Danish Pathology Register to map unique tumors taking reexcisions and changes in diagnosis into account.
}
