% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatR.R
\name{updatR}
\alias{updatR}
\title{updatR}
\usage{
updatR(data, td.frame, exclusion)
}
\arguments{
\item{data}{data frame with case IDS}

\item{td.frame}{Time-dependent frame containing CCI values to be changed}

\item{exclusion}{Vector of diagnosis codes to be excluded}
}
\value{
An updated td.frame
}
\description{
Updating of Charlson Comorbidity Index values over time#'
}
