% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatR.R
\name{updatR}
\alias{updatR}
\title{updatR}
\usage{
updatR(data, exclusion, match = "start")
}
\arguments{
\item{data}{data frame with case IDS}

\item{exclusion}{Vector of diagnosis codes to be excluded (e.g. c("DC1", "DC2", "DC3"))}

\item{lpr}{lpr data frame. Loaded automatically if missing}

\item{td.frame}{Time-dependent frame containing CCI values to be changed. Loaded automatically if missing}
}
\value{
An updated td.frame
}
\description{
Updating of Charlson Comorbidity Index values over time#'
}
